Agricultural Mechanization Services by Region

This dataset contains comprehensive statistics on agricultural mechanization services by region, including service organizations, personnel, investment, and output. The data is extracted from the China Agricultural Machinery Industry Yearbook, covering all provinces and regions in China.

Usage

MachineService

Format

A data frame:

province

character. Province name, including national total.

year

character. Year of the statistics.

chn_block4

character. Service indicator in Chinese (e.g., "Year-end Organizations", "Year-end Personnel", "Service Income").

value

numeric. The statistical value.

units

character. Units of measurement (e.g., "organizations", "persons", "10,000 yuan").

variables

character. Variable name in coded format.

Source

China Agricultural Machinery Industry Yearbook, China Agricultural Machinery Industry Association

Details

The dataset covers agricultural mechanization services from 2010 to the latest available year.

Service indicators include organization counts, personnel numbers, investment amounts, and service income.

Data structure evolved over time: separate tables from 2010-2013, merged tables from 2014-2017, and fully integrated tables from 2018-2023.

Data is in long format for easy analysis and visualization.

Examples

# View the structure of the dataset
str(MachineService)
#> 'data.frame':	8732 obs. of  6 variables:
#>  $ province  : chr  "全国" "北京" "天津" "河北" ...
#>  $ year      : chr  "2024" "2024" "2024" "2024" ...
#>  $ chn_block4: chr  "年末机构数A" "年末机构数A" "年末机构数A" "年末机构数A" ...
#>  $ value     : num  198698 250 218 5475 4646 ...
#>  $ units     : chr  "个" "个" "个" "个" ...
#>  $ variables : chr  "v9_t1_fwzz_jgs" "v9_t1_fwzz_jgs" "v9_t1_fwzz_jgs" "v9_t1_fwzz_jgs" ...
# Filter data for year-end organizations in 2023
MachineService[MachineService$year == "2023", ]
#>      province year       chn_block4       value units        variables
#> 609      全国 2023      年末机构数A   195606.00    个   v9_t1_fwzz_jgs
#> 610      北京 2023      年末机构数A      225.00    个   v9_t1_fwzz_jgs
#> 611      天津 2023      年末机构数A      214.00    个   v9_t1_fwzz_jgs
#> 612      河北 2023      年末机构数A     5533.00    个   v9_t1_fwzz_jgs
#> 613      山西 2023      年末机构数A     4625.00    个   v9_t1_fwzz_jgs
#> 614    内蒙古 2023      年末机构数A     4448.00    个   v9_t1_fwzz_jgs
#> 615      辽宁 2023      年末机构数A     5141.00    个   v9_t1_fwzz_jgs
#> 616      吉林 2023      年末机构数A     9731.00    个   v9_t1_fwzz_jgs
#> 617    黑龙江 2023      年末机构数A    22623.00    个   v9_t1_fwzz_jgs
#> 618      上海 2023      年末机构数A      309.00    个   v9_t1_fwzz_jgs
#> 619      江苏 2023      年末机构数A    12203.00    个   v9_t1_fwzz_jgs
#> 620      浙江 2023      年末机构数A     3519.00    个   v9_t1_fwzz_jgs
#> 621      安徽 2023      年末机构数A    11107.00    个   v9_t1_fwzz_jgs
#> 622      福建 2023      年末机构数A     1036.00    个   v9_t1_fwzz_jgs
#> 623      江西 2023      年末机构数A    12700.00    个   v9_t1_fwzz_jgs
#> 624      山东 2023      年末机构数A    21428.00    个   v9_t1_fwzz_jgs
#> 625      河南 2023      年末机构数A    13757.00    个   v9_t1_fwzz_jgs
#> 626      湖北 2023      年末机构数A     6744.00    个   v9_t1_fwzz_jgs
#> 627      湖南 2023      年末机构数A    19194.00    个   v9_t1_fwzz_jgs
#> 628      广东 2023      年末机构数A     2542.00    个   v9_t1_fwzz_jgs
#> 629      广西 2023      年末机构数A     3073.00    个   v9_t1_fwzz_jgs
#> 630      海南 2023      年末机构数A      181.00    个   v9_t1_fwzz_jgs
#> 631      重庆 2023      年末机构数A     4831.00    个   v9_t1_fwzz_jgs
#> 632      四川 2023      年末机构数A    16449.00    个   v9_t1_fwzz_jgs
#> 633      贵州 2023      年末机构数A     1951.00    个   v9_t1_fwzz_jgs
#> 634      云南 2023      年末机构数A     1000.00    个   v9_t1_fwzz_jgs
#> 635      西藏 2023      年末机构数A      231.00    个   v9_t1_fwzz_jgs
#> 636      陕西 2023      年末机构数A     2358.00    个   v9_t1_fwzz_jgs
#> 637      甘肃 2023      年末机构数A     4674.00    个   v9_t1_fwzz_jgs
#> 638      青海 2023      年末机构数A      853.00    个   v9_t1_fwzz_jgs
#> 639      宁夏 2023      年末机构数A      473.00    个   v9_t1_fwzz_jgs
#> 640      新疆 2023      年末机构数A     1676.00    个   v9_t1_fwzz_jgs
#> 641      全国 2023        年末人数A  2103348.00    人    v9_t1_fwzz_rs
#> 642      北京 2023        年末人数A     1541.00    人    v9_t1_fwzz_rs
#> 643      天津 2023        年末人数A     5456.00    人    v9_t1_fwzz_rs
#> 644      河北 2023        年末人数A    62072.00    人    v9_t1_fwzz_rs
#> 645      山西 2023        年末人数A    30531.00    人    v9_t1_fwzz_rs
#> 646    内蒙古 2023        年末人数A    44837.00    人    v9_t1_fwzz_rs
#> 647      辽宁 2023        年末人数A    59506.00    人    v9_t1_fwzz_rs
#> 648      吉林 2023        年末人数A    94947.00    人    v9_t1_fwzz_rs
#> 649    黑龙江 2023        年末人数A    92848.00    人    v9_t1_fwzz_rs
#> 650      上海 2023        年末人数A     2186.00    人    v9_t1_fwzz_rs
#> 651      江苏 2023        年末人数A   410864.00    人    v9_t1_fwzz_rs
#> 652      浙江 2023        年末人数A    28043.00    人    v9_t1_fwzz_rs
#> 653      安徽 2023        年末人数A   125061.00    人    v9_t1_fwzz_rs
#> 654      福建 2023        年末人数A    23205.00    人    v9_t1_fwzz_rs
#> 655      江西 2023        年末人数A    82445.00    人    v9_t1_fwzz_rs
#> 656      山东 2023        年末人数A   209603.00    人    v9_t1_fwzz_rs
#> 657      河南 2023        年末人数A   154356.00    人    v9_t1_fwzz_rs
#> 658      湖北 2023        年末人数A   148832.00    人    v9_t1_fwzz_rs
#> 659      湖南 2023        年末人数A   132682.00    人    v9_t1_fwzz_rs
#> 660      广东 2023        年末人数A    27958.00    人    v9_t1_fwzz_rs
#> 661      广西 2023        年末人数A    42799.00    人    v9_t1_fwzz_rs
#> 662      海南 2023        年末人数A     1158.00    人    v9_t1_fwzz_rs
#> 663      重庆 2023        年末人数A    86494.00    人    v9_t1_fwzz_rs
#> 664      四川 2023        年末人数A    84735.00    人    v9_t1_fwzz_rs
#> 665      贵州 2023        年末人数A    21016.00    人    v9_t1_fwzz_rs
#> 666      云南 2023        年末人数A    14836.00    人    v9_t1_fwzz_rs
#> 667      西藏 2023        年末人数A     2467.00    人    v9_t1_fwzz_rs
#> 668      陕西 2023        年末人数A    36668.00    人    v9_t1_fwzz_rs
#> 669      甘肃 2023        年末人数A    38792.00    人    v9_t1_fwzz_rs
#> 670      青海 2023        年末人数A     4976.00    人    v9_t1_fwzz_rs
#> 671      宁夏 2023        年末人数A     6700.00    人    v9_t1_fwzz_rs
#> 672      新疆 2023        年末人数A    21102.00    人    v9_t1_fwzz_rs
#> 673      全国 2023      年末机构数B    79470.00    个  v9_t1_zyhzs_jgs
#> 674      北京 2023      年末机构数B      166.00    个  v9_t1_zyhzs_jgs
#> 675      天津 2023      年末机构数B      163.00    个  v9_t1_zyhzs_jgs
#> 676      河北 2023      年末机构数B     2624.00    个  v9_t1_zyhzs_jgs
#> 677      山西 2023      年末机构数B     2777.00    个  v9_t1_zyhzs_jgs
#> 678    内蒙古 2023      年末机构数B     3220.00    个  v9_t1_zyhzs_jgs
#> 679      辽宁 2023      年末机构数B     3914.00    个  v9_t1_zyhzs_jgs
#> 680      吉林 2023      年末机构数B     6721.00    个  v9_t1_zyhzs_jgs
#> 681    黑龙江 2023      年末机构数B     2299.00    个  v9_t1_zyhzs_jgs
#> 682      上海 2023      年末机构数B      189.00    个  v9_t1_zyhzs_jgs
#> 683      江苏 2023      年末机构数B     6925.00    个  v9_t1_zyhzs_jgs
#> 684      浙江 2023      年末机构数B     1454.00    个  v9_t1_zyhzs_jgs
#> 685      安徽 2023      年末机构数B     6201.00    个  v9_t1_zyhzs_jgs
#> 686      福建 2023      年末机构数B      760.00    个  v9_t1_zyhzs_jgs
#> 687      江西 2023      年末机构数B     1663.00    个  v9_t1_zyhzs_jgs
#> 688      山东 2023      年末机构数B     9217.00    个  v9_t1_zyhzs_jgs
#> 689      河南 2023      年末机构数B     7640.00    个  v9_t1_zyhzs_jgs
#> 690      湖北 2023      年末机构数B     3588.00    个  v9_t1_zyhzs_jgs
#> 691      湖南 2023      年末机构数B     5780.00    个  v9_t1_zyhzs_jgs
#> 692      广东 2023      年末机构数B     1071.00    个  v9_t1_zyhzs_jgs
#> 693      广西 2023      年末机构数B     1691.00    个  v9_t1_zyhzs_jgs
#> 694      海南 2023      年末机构数B       88.00    个  v9_t1_zyhzs_jgs
#> 695      重庆 2023      年末机构数B     1053.00    个  v9_t1_zyhzs_jgs
#> 696      四川 2023      年末机构数B     1763.00    个  v9_t1_zyhzs_jgs
#> 697      贵州 2023      年末机构数B     1008.00    个  v9_t1_zyhzs_jgs
#> 698      云南 2023      年末机构数B      551.00    个  v9_t1_zyhzs_jgs
#> 699      西藏 2023      年末机构数B      169.00    个  v9_t1_zyhzs_jgs
#> 700      陕西 2023      年末机构数B     1487.00    个  v9_t1_zyhzs_jgs
#> 701      甘肃 2023      年末机构数B     3057.00    个  v9_t1_zyhzs_jgs
#> 702      青海 2023      年末机构数B      195.00    个  v9_t1_zyhzs_jgs
#> 703      宁夏 2023      年末机构数B      275.00    个  v9_t1_zyhzs_jgs
#> 704      新疆 2023      年末机构数B     1440.00    个  v9_t1_zyhzs_jgs
#> 705      全国 2023        年末人数B  1412352.00    人   v9_t1_zyhzs_rs
#> 706      北京 2023        年末人数B      928.00    人   v9_t1_zyhzs_rs
#> 707      天津 2023        年末人数B     4893.00    人   v9_t1_zyhzs_rs
#> 708      河北 2023        年末人数B    46925.00    人   v9_t1_zyhzs_rs
#> 709      山西 2023        年末人数B    24179.00    人   v9_t1_zyhzs_rs
#> 710    内蒙古 2023        年末人数B    34294.00    人   v9_t1_zyhzs_rs
#> 711      辽宁 2023        年末人数B    48257.00    人   v9_t1_zyhzs_rs
#> 712      吉林 2023        年末人数B    68445.00    人   v9_t1_zyhzs_rs
#> 713    黑龙江 2023        年末人数B    24231.00    人   v9_t1_zyhzs_rs
#> 714      上海 2023        年末人数B     1477.00    人   v9_t1_zyhzs_rs
#> 715      江苏 2023        年末人数B   297051.00    人   v9_t1_zyhzs_rs
#> 716      浙江 2023        年末人数B    18312.00    人   v9_t1_zyhzs_rs
#> 717      安徽 2023        年末人数B    87028.00    人   v9_t1_zyhzs_rs
#> 718      福建 2023        年末人数B    20002.00    人   v9_t1_zyhzs_rs
#> 719      江西 2023        年末人数B    27631.00    人   v9_t1_zyhzs_rs
#> 720      山东 2023        年末人数B   143471.00    人   v9_t1_zyhzs_rs
#> 721      河南 2023        年末人数B   125111.00    人   v9_t1_zyhzs_rs
#> 722      湖北 2023        年末人数B   110230.00    人   v9_t1_zyhzs_rs
#> 723      湖南 2023        年末人数B    77240.00    人   v9_t1_zyhzs_rs
#> 724      广东 2023        年末人数B    19508.00    人   v9_t1_zyhzs_rs
#> 725      广西 2023        年末人数B    25100.00    人   v9_t1_zyhzs_rs
#> 726      海南 2023        年末人数B      798.00    人   v9_t1_zyhzs_rs
#> 727      重庆 2023        年末人数B    56949.00    人   v9_t1_zyhzs_rs
#> 728      四川 2023        年末人数B    34693.00    人   v9_t1_zyhzs_rs
#> 729      贵州 2023        年末人数B    10533.00    人   v9_t1_zyhzs_rs
#> 730      云南 2023        年末人数B    11801.00    人   v9_t1_zyhzs_rs
#> 731      西藏 2023        年末人数B     2301.00    人   v9_t1_zyhzs_rs
#> 732      陕西 2023        年末人数B    29798.00    人   v9_t1_zyhzs_rs
#> 733      甘肃 2023        年末人数B    30732.00    人   v9_t1_zyhzs_rs
#> 734      青海 2023        年末人数B     3069.00    人   v9_t1_zyhzs_rs
#> 735      宁夏 2023        年末人数B     3933.00    人   v9_t1_zyhzs_rs
#> 736      新疆 2023        年末人数B    19113.00    人   v9_t1_zyhzs_rs
#> 737      全国 2023      年末机构数C 39193020.00    个    v9_t1_njh_jgs
#> 738      北京 2023      年末机构数C    11788.00    个    v9_t1_njh_jgs
#> 739      天津 2023      年末机构数C    34668.00    个    v9_t1_njh_jgs
#> 740      河北 2023      年末机构数C  2420982.00    个    v9_t1_njh_jgs
#> 741      山西 2023      年末机构数C   510702.00    个    v9_t1_njh_jgs
#> 742    内蒙古 2023      年末机构数C  1248251.00    个    v9_t1_njh_jgs
#> 743      辽宁 2023      年末机构数C   536991.00    个    v9_t1_njh_jgs
#> 744      吉林 2023      年末机构数C  1100789.00    个    v9_t1_njh_jgs
#> 745    黑龙江 2023      年末机构数C  1169734.00    个    v9_t1_njh_jgs
#> 746      上海 2023      年末机构数C     4462.00    个    v9_t1_njh_jgs
#> 747      江苏 2023      年末机构数C  1038229.00    个    v9_t1_njh_jgs
#> 748      浙江 2023      年末机构数C   466712.00    个    v9_t1_njh_jgs
#> 749      安徽 2023      年末机构数C  3117163.00    个    v9_t1_njh_jgs
#> 750      福建 2023      年末机构数C   600506.00    个    v9_t1_njh_jgs
#> 751      江西 2023      年末机构数C  1023125.00    个    v9_t1_njh_jgs
#> 752      山东 2023      年末机构数C  3914821.00    个    v9_t1_njh_jgs
#> 753      河南 2023      年末机构数C  4890000.00    个    v9_t1_njh_jgs
#> 754      湖北 2023      年末机构数C  1724808.00    个    v9_t1_njh_jgs
#> 755      湖南 2023      年末机构数C  2046751.00    个    v9_t1_njh_jgs
#> 756      广东 2023      年末机构数C  1042910.00    个    v9_t1_njh_jgs
#> 757      广西 2023      年末机构数C  2468660.00    个    v9_t1_njh_jgs
#> 758      海南 2023      年末机构数C   208263.00    个    v9_t1_njh_jgs
#> 759      重庆 2023      年末机构数C  1212476.00    个    v9_t1_njh_jgs
#> 760      四川 2023      年末机构数C  2394351.00    个    v9_t1_njh_jgs
#> 761      贵州 2023      年末机构数C   572674.00    个    v9_t1_njh_jgs
#> 762      云南 2023      年末机构数C  1969794.00    个    v9_t1_njh_jgs
#> 763      西藏 2023      年末机构数C   228953.00    个    v9_t1_njh_jgs
#> 764      陕西 2023      年末机构数C   956154.00    个    v9_t1_njh_jgs
#> 765      甘肃 2023      年末机构数C  1211975.00    个    v9_t1_njh_jgs
#> 766      青海 2023      年末机构数C   196380.00    个    v9_t1_njh_jgs
#> 767      宁夏 2023      年末机构数C   252033.00    个    v9_t1_njh_jgs
#> 768      新疆 2023      年末机构数C   590416.00    个    v9_t1_njh_jgs
#> 769      全国 2023        年末人数C 46615424.00    人     v9_t1_njh_rs
#> 770      北京 2023        年末人数C    12561.00    人     v9_t1_njh_rs
#> 771      天津 2023        年末人数C    41449.00    人     v9_t1_njh_rs
#> 772      河北 2023        年末人数C  2808740.00    人     v9_t1_njh_rs
#> 773      山西 2023        年末人数C   648950.00    人     v9_t1_njh_rs
#> 774    内蒙古 2023        年末人数C  1587289.00    人     v9_t1_njh_rs
#> 775      辽宁 2023        年末人数C   651342.00    人     v9_t1_njh_rs
#> 776      吉林 2023        年末人数C  1305715.00    人     v9_t1_njh_rs
#> 777    黑龙江 2023        年末人数C  1458849.00    人     v9_t1_njh_rs
#> 778      上海 2023        年末人数C     6036.00    人     v9_t1_njh_rs
#> 779      江苏 2023        年末人数C  1242430.00    人     v9_t1_njh_rs
#> 780      浙江 2023        年末人数C   539734.00    人     v9_t1_njh_rs
#> 781      安徽 2023        年末人数C  3651851.00    人     v9_t1_njh_rs
#> 782      福建 2023        年末人数C   672279.00    人     v9_t1_njh_rs
#> 783      江西 2023        年末人数C  1401998.00    人     v9_t1_njh_rs
#> 784      山东 2023        年末人数C  4545244.00    人     v9_t1_njh_rs
#> 785      河南 2023        年末人数C  5489074.00    人     v9_t1_njh_rs
#> 786      湖北 2023        年末人数C  2239058.00    人     v9_t1_njh_rs
#> 787      湖南 2023        年末人数C  2716203.00    人     v9_t1_njh_rs
#> 788      广东 2023        年末人数C  1198857.00    人     v9_t1_njh_rs
#> 789      广西 2023        年末人数C  3153394.00    人     v9_t1_njh_rs
#> 790      海南 2023        年末人数C   239341.00    人     v9_t1_njh_rs
#> 791      重庆 2023        年末人数C  1324445.00    人     v9_t1_njh_rs
#> 792      四川 2023        年末人数C  2802935.00    人     v9_t1_njh_rs
#> 793      贵州 2023        年末人数C   818626.00    人     v9_t1_njh_rs
#> 794      云南 2023        年末人数C  2099837.00    人     v9_t1_njh_rs
#> 795      西藏 2023        年末人数C   268726.00    人     v9_t1_njh_rs
#> 796      陕西 2023        年末人数C  1100590.00    人     v9_t1_njh_rs
#> 797      甘肃 2023        年末人数C  1325238.00    人     v9_t1_njh_rs
#> 798      青海 2023        年末人数C   219044.00    人     v9_t1_njh_rs
#> 799      宁夏 2023        年末人数C   354965.00    人     v9_t1_njh_rs
#> 800      新疆 2023        年末人数C   649860.00    人     v9_t1_njh_rs
#> 801      全国 2023      年末机构数D  4086769.00    个    v9_t1_zyh_jgs
#> 802      北京 2023      年末机构数D      512.00    个    v9_t1_zyh_jgs
#> 803      天津 2023      年末机构数D     7784.00    个    v9_t1_zyh_jgs
#> 804      河北 2023      年末机构数D   179213.00    个    v9_t1_zyh_jgs
#> 805      山西 2023      年末机构数D    70004.00    个    v9_t1_zyh_jgs
#> 806    内蒙古 2023      年末机构数D   108438.00    个    v9_t1_zyh_jgs
#> 807      辽宁 2023      年末机构数D    59143.00    个    v9_t1_zyh_jgs
#> 808      吉林 2023      年末机构数D    29654.00    个    v9_t1_zyh_jgs
#> 809    黑龙江 2023      年末机构数D   135036.00    个    v9_t1_zyh_jgs
#> 810      上海 2023      年末机构数D      940.00    个    v9_t1_zyh_jgs
#> 811      江苏 2023      年末机构数D   176342.00    个    v9_t1_zyh_jgs
#> 812      浙江 2023      年末机构数D    81312.00    个    v9_t1_zyh_jgs
#> 813      安徽 2023      年末机构数D   417083.00    个    v9_t1_zyh_jgs
#> 814      福建 2023      年末机构数D    89956.00    个    v9_t1_zyh_jgs
#> 815      江西 2023      年末机构数D   330383.00    个    v9_t1_zyh_jgs
#> 816      山东 2023      年末机构数D   560044.00    个    v9_t1_zyh_jgs
#> 817      河南 2023      年末机构数D   232377.00    个    v9_t1_zyh_jgs
#> 818      湖北 2023      年末机构数D   166335.00    个    v9_t1_zyh_jgs
#> 819      湖南 2023      年末机构数D   180452.00    个    v9_t1_zyh_jgs
#> 820      广东 2023      年末机构数D    51114.00    个    v9_t1_zyh_jgs
#> 821      广西 2023      年末机构数D    29625.00    个    v9_t1_zyh_jgs
#> 822      海南 2023      年末机构数D    47100.00    个    v9_t1_zyh_jgs
#> 823      重庆 2023      年末机构数D   140610.00    个    v9_t1_zyh_jgs
#> 824      四川 2023      年末机构数D   189595.00    个    v9_t1_zyh_jgs
#> 825      贵州 2023      年末机构数D   157206.00    个    v9_t1_zyh_jgs
#> 826      云南 2023      年末机构数D    61343.00    个    v9_t1_zyh_jgs
#> 827      西藏 2023      年末机构数D      183.00    个    v9_t1_zyh_jgs
#> 828      陕西 2023      年末机构数D   222268.00    个    v9_t1_zyh_jgs
#> 829      甘肃 2023      年末机构数D   191079.00    个    v9_t1_zyh_jgs
#> 830      青海 2023      年末机构数D     4286.00    个    v9_t1_zyh_jgs
#> 831      宁夏 2023      年末机构数D    37906.00    个    v9_t1_zyh_jgs
#> 832      新疆 2023      年末机构数D   121330.00    个    v9_t1_zyh_jgs
#> 833      全国 2023        年末人数D  5732807.00    人     v9_t1_zyh_rs
#> 834      北京 2023        年末人数D      769.00    人     v9_t1_zyh_rs
#> 835      天津 2023        年末人数D    11129.00    人     v9_t1_zyh_rs
#> 836      河北 2023        年末人数D   277096.00    人     v9_t1_zyh_rs
#> 837      山西 2023        年末人数D    94591.00    人     v9_t1_zyh_rs
#> 838    内蒙古 2023        年末人数D   155893.00    人     v9_t1_zyh_rs
#> 839      辽宁 2023        年末人数D    86064.00    人     v9_t1_zyh_rs
#> 840      吉林 2023        年末人数D    32335.00    人     v9_t1_zyh_rs
#> 841    黑龙江 2023        年末人数D   195571.00    人     v9_t1_zyh_rs
#> 842      上海 2023        年末人数D     1428.00    人     v9_t1_zyh_rs
#> 843      江苏 2023        年末人数D   264692.00    人     v9_t1_zyh_rs
#> 844      浙江 2023        年末人数D    95903.00    人     v9_t1_zyh_rs
#> 845      安徽 2023        年末人数D   586861.00    人     v9_t1_zyh_rs
#> 846      福建 2023        年末人数D   110358.00    人     v9_t1_zyh_rs
#> 847      江西 2023        年末人数D   495339.00    人     v9_t1_zyh_rs
#> 848      山东 2023        年末人数D   782028.00    人     v9_t1_zyh_rs
#> 849      河南 2023        年末人数D   372472.00    人     v9_t1_zyh_rs
#> 850      湖北 2023        年末人数D   325121.00    人     v9_t1_zyh_rs
#> 851      湖南 2023        年末人数D   270372.00    人     v9_t1_zyh_rs
#> 852      广东 2023        年末人数D    82531.00    人     v9_t1_zyh_rs
#> 853      广西 2023        年末人数D    68852.00    人     v9_t1_zyh_rs
#> 854      海南 2023        年末人数D    51173.00    人     v9_t1_zyh_rs
#> 855      重庆 2023        年末人数D   167567.00    人     v9_t1_zyh_rs
#> 856      四川 2023        年末人数D   242128.00    人     v9_t1_zyh_rs
#> 857      贵州 2023        年末人数D   209929.00    人     v9_t1_zyh_rs
#> 858      云南 2023        年末人数D    70021.00    人     v9_t1_zyh_rs
#> 859      西藏 2023        年末人数D      268.00    人     v9_t1_zyh_rs
#> 860      陕西 2023        年末人数D   244540.00    人     v9_t1_zyh_rs
#> 861      甘肃 2023        年末人数D   221861.00    人     v9_t1_zyh_rs
#> 862      青海 2023        年末人数D     5937.00    人     v9_t1_zyh_rs
#> 863      宁夏 2023        年末人数D    41913.00    人     v9_t1_zyh_rs
#> 864      新疆 2023        年末人数D   155313.00    人     v9_t1_zyh_rs
#> 865      全国 2023        年末人数G 49928434.00    人  v9_t1_njcyry_rs
#> 866      北京 2023        年末人数G    17604.00    人  v9_t1_njcyry_rs
#> 867      天津 2023        年末人数G    60986.00    人  v9_t1_njcyry_rs
#> 868      河北 2023        年末人数G  3268079.00    人  v9_t1_njcyry_rs
#> 869      山西 2023        年末人数G   704283.00    人  v9_t1_njcyry_rs
#> 870    内蒙古 2023        年末人数G  1797205.00    人  v9_t1_njcyry_rs
#> 871      辽宁 2023        年末人数G   831064.00    人  v9_t1_njcyry_rs
#> 872      吉林 2023        年末人数G  1353191.00    人  v9_t1_njcyry_rs
#> 873    黑龙江 2023        年末人数G  1741918.00    人  v9_t1_njcyry_rs
#> 874      上海 2023        年末人数G    14136.00    人  v9_t1_njcyry_rs
#> 875      江苏 2023        年末人数G  1559645.00    人  v9_t1_njcyry_rs
#> 876      浙江 2023        年末人数G   546315.00    人  v9_t1_njcyry_rs
#> 877      安徽 2023        年末人数G  3840352.00    人  v9_t1_njcyry_rs
#> 878      福建 2023        年末人数G   698424.00    人  v9_t1_njcyry_rs
#> 879      江西 2023        年末人数G  1403260.00    人  v9_t1_njcyry_rs
#> 880      山东 2023        年末人数G  4675254.00    人  v9_t1_njcyry_rs
#> 881      河南 2023        年末人数G  5774998.00    人  v9_t1_njcyry_rs
#> 882      湖北 2023        年末人数G  2249249.00    人  v9_t1_njcyry_rs
#> 883      湖南 2023        年末人数G  2799088.00    人  v9_t1_njcyry_rs
#> 884      广东 2023        年末人数G  1226209.00    人  v9_t1_njcyry_rs
#> 885      广西 2023        年末人数G  3295941.00    人  v9_t1_njcyry_rs
#> 886      海南 2023        年末人数G   243491.00    人  v9_t1_njcyry_rs
#> 887      重庆 2023        年末人数G  1377979.00    人  v9_t1_njcyry_rs
#> 888      四川 2023        年末人数G  2841383.00    人  v9_t1_njcyry_rs
#> 889      贵州 2023        年末人数G  1114734.00    人  v9_t1_njcyry_rs
#> 890      云南 2023        年末人数G  2251208.00    人  v9_t1_njcyry_rs
#> 891      西藏 2023        年末人数G   321149.00    人  v9_t1_njcyry_rs
#> 892      陕西 2023        年末人数G  1104253.00    人  v9_t1_njcyry_rs
#> 893      甘肃 2023        年末人数G  1393248.00    人  v9_t1_njcyry_rs
#> 894      青海 2023        年末人数G   268423.00    人  v9_t1_njcyry_rs
#> 895      宁夏 2023        年末人数G   359575.00    人  v9_t1_njcyry_rs
#> 896      新疆 2023        年末人数G   690707.00    人  v9_t1_njcyry_rs
#> 897      全国 2023      年末机构数F   142782.00    个   v9_t1_njwx_jgs
#> 898      北京 2023      年末机构数F      138.00    个   v9_t1_njwx_jgs
#> 899      天津 2023      年末机构数F      183.00    个   v9_t1_njwx_jgs
#> 900      河北 2023      年末机构数F    11713.00    个   v9_t1_njwx_jgs
#> 901      山西 2023      年末机构数F     4809.00    个   v9_t1_njwx_jgs
#> 902    内蒙古 2023      年末机构数F     6481.00    个   v9_t1_njwx_jgs
#> 903      辽宁 2023      年末机构数F     3158.00    个   v9_t1_njwx_jgs
#> 904      吉林 2023      年末机构数F     6342.00    个   v9_t1_njwx_jgs
#> 905    黑龙江 2023      年末机构数F     6545.00    个   v9_t1_njwx_jgs
#> 906      上海 2023      年末机构数F       25.00    个   v9_t1_njwx_jgs
#> 907      江苏 2023      年末机构数F     3145.00    个   v9_t1_njwx_jgs
#> 908      浙江 2023      年末机构数F     1830.00    个   v9_t1_njwx_jgs
#> 909      安徽 2023      年末机构数F     9116.00    个   v9_t1_njwx_jgs
#> 910      福建 2023      年末机构数F      291.00    个   v9_t1_njwx_jgs
#> 911      江西 2023      年末机构数F     9036.00    个   v9_t1_njwx_jgs
#> 912      山东 2023      年末机构数F     4031.00    个   v9_t1_njwx_jgs
#> 913      河南 2023      年末机构数F    18279.00    个   v9_t1_njwx_jgs
#> 914      湖北 2023      年末机构数F     5139.00    个   v9_t1_njwx_jgs
#> 915      湖南 2023      年末机构数F     5045.00    个   v9_t1_njwx_jgs
#> 916      广东 2023      年末机构数F     6576.00    个   v9_t1_njwx_jgs
#> 917      广西 2023      年末机构数F     3587.00    个   v9_t1_njwx_jgs
#> 918      海南 2023      年末机构数F      736.00    个   v9_t1_njwx_jgs
#> 919      重庆 2023      年末机构数F     2529.00    个   v9_t1_njwx_jgs
#> 920      四川 2023      年末机构数F     7531.00    个   v9_t1_njwx_jgs
#> 921      贵州 2023      年末机构数F     4342.00    个   v9_t1_njwx_jgs
#> 922      云南 2023      年末机构数F     8535.00    个   v9_t1_njwx_jgs
#> 923      西藏 2023      年末机构数F      234.00    个   v9_t1_njwx_jgs
#> 924      陕西 2023      年末机构数F     3494.00    个   v9_t1_njwx_jgs
#> 925      甘肃 2023      年末机构数F     4722.00    个   v9_t1_njwx_jgs
#> 926      青海 2023      年末机构数F     1042.00    个   v9_t1_njwx_jgs
#> 927      宁夏 2023      年末机构数F     1713.00    个   v9_t1_njwx_jgs
#> 928      新疆 2023      年末机构数F     2057.00    个   v9_t1_njwx_jgs
#> 929      全国 2023        年末人数F   355285.00    人    v9_t1_njwx_rs
#> 930      北京 2023        年末人数F      278.00    人    v9_t1_njwx_rs
#> 931      天津 2023        年末人数F      387.00    人    v9_t1_njwx_rs
#> 932      河北 2023        年末人数F    28873.00    人    v9_t1_njwx_rs
#> 933      山西 2023        年末人数F    12012.00    人    v9_t1_njwx_rs
#> 934    内蒙古 2023        年末人数F    13664.00    人    v9_t1_njwx_rs
#> 935      辽宁 2023        年末人数F     7272.00    人    v9_t1_njwx_rs
#> 936      吉林 2023        年末人数F    14669.00    人    v9_t1_njwx_rs
#> 937    黑龙江 2023        年末人数F    20638.00    人    v9_t1_njwx_rs
#> 938      上海 2023        年末人数F       93.00    人    v9_t1_njwx_rs
#> 939      江苏 2023        年末人数F    10279.00    人    v9_t1_njwx_rs
#> 940      浙江 2023        年末人数F     3376.00    人    v9_t1_njwx_rs
#> 941      安徽 2023        年末人数F    21792.00    人    v9_t1_njwx_rs
#> 942      福建 2023        年末人数F      531.00    人    v9_t1_njwx_rs
#> 943      江西 2023        年末人数F    28113.00    人    v9_t1_njwx_rs
#> 944      山东 2023        年末人数F    11936.00    人    v9_t1_njwx_rs
#> 945      河南 2023        年末人数F    38233.00    人    v9_t1_njwx_rs
#> 946      湖北 2023        年末人数F    15925.00    人    v9_t1_njwx_rs
#> 947      湖南 2023        年末人数F    13864.00    人    v9_t1_njwx_rs
#> 948      广东 2023        年末人数F    16901.00    人    v9_t1_njwx_rs
#> 949      广西 2023        年末人数F     8789.00    人    v9_t1_njwx_rs
#> 950      海南 2023        年末人数F     1818.00    人    v9_t1_njwx_rs
#> 951      重庆 2023        年末人数F     7229.00    人    v9_t1_njwx_rs
#> 952      四川 2023        年末人数F    22965.00    人    v9_t1_njwx_rs
#> 953      贵州 2023        年末人数F    11221.00    人    v9_t1_njwx_rs
#> 954      云南 2023        年末人数F    18536.00    人    v9_t1_njwx_rs
#> 955      西藏 2023        年末人数F      722.00    人    v9_t1_njwx_rs
#> 956      陕西 2023        年末人数F     7252.00    人    v9_t1_njwx_rs
#> 957      甘肃 2023        年末人数F     8135.00    人    v9_t1_njwx_rs
#> 958      青海 2023        年末人数F     1638.00    人    v9_t1_njwx_rs
#> 959      宁夏 2023        年末人数F     3147.00    人    v9_t1_njwx_rs
#> 960      新疆 2023        年末人数F     4043.00    人    v9_t1_njwx_rs
#> 961      全国 2023     财政资金投入   933466.46  万元  v9_t1_cztr_czzj
#> 962      北京 2023     财政资金投入     5836.71  万元  v9_t1_cztr_czzj
#> 963      天津 2023     财政资金投入     5177.81  万元  v9_t1_cztr_czzj
#> 964      河北 2023     财政资金投入    14183.60  万元  v9_t1_cztr_czzj
#> 965      山西 2023     财政资金投入    75801.00  万元  v9_t1_cztr_czzj
#> 966    内蒙古 2023     财政资金投入   105765.96  万元  v9_t1_cztr_czzj
#> 967      辽宁 2023     财政资金投入    49133.76  万元  v9_t1_cztr_czzj
#> 968      吉林 2023     财政资金投入    70608.16  万元  v9_t1_cztr_czzj
#> 969    黑龙江 2023     财政资金投入    11717.38  万元  v9_t1_cztr_czzj
#> 970      上海 2023     财政资金投入    10236.00  万元  v9_t1_cztr_czzj
#> 971      江苏 2023     财政资金投入   163624.23  万元  v9_t1_cztr_czzj
#> 972      浙江 2023     财政资金投入    16941.13  万元  v9_t1_cztr_czzj
#> 973      安徽 2023     财政资金投入    23711.57  万元  v9_t1_cztr_czzj
#> 974      福建 2023     财政资金投入     3797.57  万元  v9_t1_cztr_czzj
#> 975      江西 2023     财政资金投入    17675.00  万元  v9_t1_cztr_czzj
#> 976      山东 2023     财政资金投入    27693.65  万元  v9_t1_cztr_czzj
#> 977      河南 2023     财政资金投入    42346.90  万元  v9_t1_cztr_czzj
#> 978      湖北 2023     财政资金投入    59042.00  万元  v9_t1_cztr_czzj
#> 979      湖南 2023     财政资金投入    62593.24  万元  v9_t1_cztr_czzj
#> 980      广东 2023     财政资金投入     9247.99  万元  v9_t1_cztr_czzj
#> 981      广西 2023     财政资金投入     6133.49  万元  v9_t1_cztr_czzj
#> 982      海南 2023     财政资金投入     3965.21  万元  v9_t1_cztr_czzj
#> 983      重庆 2023     财政资金投入     9456.02  万元  v9_t1_cztr_czzj
#> 984      四川 2023     财政资金投入    36042.00  万元  v9_t1_cztr_czzj
#> 985      贵州 2023     财政资金投入    19299.73  万元  v9_t1_cztr_czzj
#> 986      云南 2023     财政资金投入     6766.75  万元  v9_t1_cztr_czzj
#> 987      西藏 2023     财政资金投入        0.01  万元  v9_t1_cztr_czzj
#> 988      陕西 2023     财政资金投入    27212.09  万元  v9_t1_cztr_czzj
#> 989      甘肃 2023     财政资金投入    19644.00  万元  v9_t1_cztr_czzj
#> 990      青海 2023     财政资金投入     3582.10  万元  v9_t1_cztr_czzj
#> 991      宁夏 2023     财政资金投入    10793.00  万元  v9_t1_cztr_czzj
#> 992      新疆 2023     财政资金投入    12962.91  万元  v9_t1_cztr_czzj
#> 993      全国 2023         科研投入    63553.55  万元  v9_t1_cztr_kytr
#> 994      北京 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 995      天津 2023         科研投入       82.81  万元  v9_t1_cztr_kytr
#> 996      河北 2023         科研投入      592.30  万元  v9_t1_cztr_kytr
#> 997      山西 2023         科研投入      485.00  万元  v9_t1_cztr_kytr
#> 998    内蒙古 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 999      辽宁 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 1000     吉林 2023         科研投入      595.57  万元  v9_t1_cztr_kytr
#> 1001   黑龙江 2023         科研投入     3050.00  万元  v9_t1_cztr_kytr
#> 1002     上海 2023         科研投入      744.00  万元  v9_t1_cztr_kytr
#> 1003     江苏 2023         科研投入     6143.00  万元  v9_t1_cztr_kytr
#> 1004     浙江 2023         科研投入     1812.60  万元  v9_t1_cztr_kytr
#> 1005     安徽 2023         科研投入     1904.60  万元  v9_t1_cztr_kytr
#> 1006     福建 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 1007     江西 2023         科研投入     2486.00  万元  v9_t1_cztr_kytr
#> 1008     山东 2023         科研投入    10953.77  万元  v9_t1_cztr_kytr
#> 1009     河南 2023         科研投入     2090.00  万元  v9_t1_cztr_kytr
#> 1010     湖北 2023         科研投入     2783.00  万元  v9_t1_cztr_kytr
#> 1011     湖南 2023         科研投入    11295.00  万元  v9_t1_cztr_kytr
#> 1012     广东 2023         科研投入     3469.30  万元  v9_t1_cztr_kytr
#> 1013     广西 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 1014     海南 2023         科研投入     1698.00  万元  v9_t1_cztr_kytr
#> 1015     重庆 2023         科研投入     1680.00  万元  v9_t1_cztr_kytr
#> 1016     四川 2023         科研投入     2111.10  万元  v9_t1_cztr_kytr
#> 1017     贵州 2023         科研投入      296.00  万元  v9_t1_cztr_kytr
#> 1018     云南 2023         科研投入       20.00  万元  v9_t1_cztr_kytr
#> 1019     西藏 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 1020     陕西 2023         科研投入      152.00  万元  v9_t1_cztr_kytr
#> 1021     甘肃 2023         科研投入     1157.00  万元  v9_t1_cztr_kytr
#> 1022     青海 2023         科研投入       30.00  万元  v9_t1_cztr_kytr
#> 1023     宁夏 2023         科研投入       34.00  万元  v9_t1_cztr_kytr
#> 1024     新疆 2023         科研投入     6880.50  万元  v9_t1_cztr_kytr
#> 1025     全国 2023         推广投入   481547.21  万元  v9_t1_cztr_tgtr
#> 1026     北京 2023         推广投入     5768.81  万元  v9_t1_cztr_tgtr
#> 1027     天津 2023         推广投入       88.30  万元  v9_t1_cztr_tgtr
#> 1028     河北 2023         推广投入     4920.15  万元  v9_t1_cztr_tgtr
#> 1029     山西 2023         推广投入     9585.00  万元  v9_t1_cztr_tgtr
#> 1030   内蒙古 2023         推广投入   104431.50  万元  v9_t1_cztr_tgtr
#> 1031     辽宁 2023         推广投入    16568.57  万元  v9_t1_cztr_tgtr
#> 1032     吉林 2023         推广投入    69584.17  万元  v9_t1_cztr_tgtr
#> 1033   黑龙江 2023         推广投入      897.46  万元  v9_t1_cztr_tgtr
#> 1034     上海 2023         推广投入     8684.00  万元  v9_t1_cztr_tgtr
#> 1035     江苏 2023         推广投入   101224.28  万元  v9_t1_cztr_tgtr
#> 1036     浙江 2023         推广投入    14493.08  万元  v9_t1_cztr_tgtr
#> 1037     安徽 2023         推广投入    10175.65  万元  v9_t1_cztr_tgtr
#> 1038     福建 2023         推广投入     2856.82  万元  v9_t1_cztr_tgtr
#> 1039     江西 2023         推广投入     9940.00  万元  v9_t1_cztr_tgtr
#> 1040     山东 2023         推广投入     9815.14  万元  v9_t1_cztr_tgtr
#> 1041     河南 2023         推广投入     2989.00  万元  v9_t1_cztr_tgtr
#> 1042     湖北 2023         推广投入    11065.00  万元  v9_t1_cztr_tgtr
#> 1043     湖南 2023         推广投入    15744.31  万元  v9_t1_cztr_tgtr
#> 1044     广东 2023         推广投入     3651.37  万元  v9_t1_cztr_tgtr
#> 1045     广西 2023         推广投入     5539.50  万元  v9_t1_cztr_tgtr
#> 1046     海南 2023         推广投入      960.00  万元  v9_t1_cztr_tgtr
#> 1047     重庆 2023         推广投入     4791.00  万元  v9_t1_cztr_tgtr
#> 1048     四川 2023         推广投入    11544.00  万元  v9_t1_cztr_tgtr
#> 1049     贵州 2023         推广投入    15228.03  万元  v9_t1_cztr_tgtr
#> 1050     云南 2023         推广投入     6328.25  万元  v9_t1_cztr_tgtr
#> 1051     西藏 2023         推广投入        0.01  万元  v9_t1_cztr_tgtr
#> 1052     陕西 2023         推广投入    12997.06  万元  v9_t1_cztr_tgtr
#> 1053     甘肃 2023         推广投入    11960.00  万元  v9_t1_cztr_tgtr
#> 1054     青海 2023         推广投入     1540.10  万元  v9_t1_cztr_tgtr
#> 1055     宁夏 2023         推广投入     4333.00  万元  v9_t1_cztr_tgtr
#> 1056     新疆 2023         推广投入     2390.16  万元  v9_t1_cztr_tgtr
#> 1057     全国 2023     基本建设投入  1240244.34  万元  v9_t1_jstr_jbjs
#> 1058     北京 2023     基本建设投入        5.00  万元  v9_t1_jstr_jbjs
#> 1059     天津 2023     基本建设投入     7357.00  万元  v9_t1_jstr_jbjs
#> 1060     河北 2023     基本建设投入    30555.40  万元  v9_t1_jstr_jbjs
#> 1061     山西 2023     基本建设投入     7882.00  万元  v9_t1_jstr_jbjs
#> 1062   内蒙古 2023     基本建设投入    85172.00  万元  v9_t1_jstr_jbjs
#> 1063     辽宁 2023     基本建设投入     2978.00  万元  v9_t1_jstr_jbjs
#> 1064     吉林 2023     基本建设投入    18157.00  万元  v9_t1_jstr_jbjs
#> 1065   黑龙江 2023     基本建设投入    11839.00  万元  v9_t1_jstr_jbjs
#> 1066     上海 2023     基本建设投入      314.00  万元  v9_t1_jstr_jbjs
#> 1067     江苏 2023     基本建设投入   179601.04  万元  v9_t1_jstr_jbjs
#> 1068     浙江 2023     基本建设投入   129881.44  万元  v9_t1_jstr_jbjs
#> 1069     安徽 2023     基本建设投入    51353.40  万元  v9_t1_jstr_jbjs
#> 1070     福建 2023     基本建设投入      203.48  万元  v9_t1_jstr_jbjs
#> 1071     江西 2023     基本建设投入    20503.00  万元  v9_t1_jstr_jbjs
#> 1072     山东 2023     基本建设投入   193284.26  万元  v9_t1_jstr_jbjs
#> 1073     河南 2023     基本建设投入     5547.00  万元  v9_t1_jstr_jbjs
#> 1074     湖北 2023     基本建设投入    19030.00  万元  v9_t1_jstr_jbjs
#> 1075     湖南 2023     基本建设投入    26346.00  万元  v9_t1_jstr_jbjs
#> 1076     广东 2023     基本建设投入    41268.47  万元  v9_t1_jstr_jbjs
#> 1077     广西 2023     基本建设投入       77.20  万元  v9_t1_jstr_jbjs
#> 1078     海南 2023     基本建设投入     1319.50  万元  v9_t1_jstr_jbjs
#> 1079     重庆 2023     基本建设投入   105507.00  万元  v9_t1_jstr_jbjs
#> 1080     四川 2023     基本建设投入   227148.00  万元  v9_t1_jstr_jbjs
#> 1081     贵州 2023     基本建设投入    24612.00  万元  v9_t1_jstr_jbjs
#> 1082     云南 2023     基本建设投入     5926.00  万元  v9_t1_jstr_jbjs
#> 1083     西藏 2023     基本建设投入      220.00  万元  v9_t1_jstr_jbjs
#> 1084     陕西 2023     基本建设投入     3635.35  万元  v9_t1_jstr_jbjs
#> 1085     甘肃 2023     基本建设投入     4831.00  万元  v9_t1_jstr_jbjs
#> 1086     青海 2023     基本建设投入    27131.00  万元  v9_t1_jstr_jbjs
#> 1087     宁夏 2023     基本建设投入     2701.00  万元  v9_t1_jstr_jbjs
#> 1088     新疆 2023     基本建设投入      270.50  万元  v9_t1_jstr_jbjs
#> 1089     全国 2023 农业机械购置投入 10160792.41  万元  v9_t1_gztr_jxgz
#> 1090     北京 2023 农业机械购置投入    16501.23  万元  v9_t1_gztr_jxgz
#> 1091     天津 2023 农业机械购置投入    54363.25  万元  v9_t1_gztr_jxgz
#> 1092     河北 2023 农业机械购置投入   636039.78  万元  v9_t1_gztr_jxgz
#> 1093     山西 2023 农业机械购置投入   202509.00  万元  v9_t1_gztr_jxgz
#> 1094   内蒙古 2023 农业机械购置投入   716631.15  万元  v9_t1_gztr_jxgz
#> 1095     辽宁 2023 农业机械购置投入   384319.01  万元  v9_t1_gztr_jxgz
#> 1096     吉林 2023 农业机械购置投入   424919.45  万元  v9_t1_gztr_jxgz
#> 1097   黑龙江 2023 农业机械购置投入   794509.90  万元  v9_t1_gztr_jxgz
#> 1098     上海 2023 农业机械购置投入    20685.00  万元  v9_t1_gztr_jxgz
#> 1099     江苏 2023 农业机械购置投入   721021.65  万元  v9_t1_gztr_jxgz
#> 1100     浙江 2023 农业机械购置投入   193271.81  万元  v9_t1_gztr_jxgz
#> 1101     安徽 2023 农业机械购置投入   793376.22  万元  v9_t1_gztr_jxgz
#> 1102     福建 2023 农业机械购置投入   109221.17  万元  v9_t1_gztr_jxgz
#> 1103     江西 2023 农业机械购置投入   283044.00  万元  v9_t1_gztr_jxgz
#> 1104     山东 2023 农业机械购置投入   820324.96  万元  v9_t1_gztr_jxgz
#> 1105     河南 2023 农业机械购置投入   624436.00  万元  v9_t1_gztr_jxgz
#> 1106     湖北 2023 农业机械购置投入   736003.40  万元  v9_t1_gztr_jxgz
#> 1107     湖南 2023 农业机械购置投入   228260.15  万元  v9_t1_gztr_jxgz
#> 1108     广东 2023 农业机械购置投入    77799.31  万元  v9_t1_gztr_jxgz
#> 1109     广西 2023 农业机械购置投入   220386.05  万元  v9_t1_gztr_jxgz
#> 1110     海南 2023 农业机械购置投入    14046.89  万元  v9_t1_gztr_jxgz
#> 1111     重庆 2023 农业机械购置投入    90215.00  万元  v9_t1_gztr_jxgz
#> 1112     四川 2023 农业机械购置投入   176705.00  万元  v9_t1_gztr_jxgz
#> 1113     贵州 2023 农业机械购置投入    68756.51  万元  v9_t1_gztr_jxgz
#> 1114     云南 2023 农业机械购置投入   215393.10  万元  v9_t1_gztr_jxgz
#> 1115     西藏 2023 农业机械购置投入    42390.82  万元  v9_t1_gztr_jxgz
#> 1116     陕西 2023 农业机械购置投入   162953.54  万元  v9_t1_gztr_jxgz
#> 1117     甘肃 2023 农业机械购置投入   270174.30  万元  v9_t1_gztr_jxgz
#> 1118     青海 2023 农业机械购置投入    37011.00  万元  v9_t1_gztr_jxgz
#> 1119     宁夏 2023 农业机械购置投入   110146.00  万元  v9_t1_gztr_jxgz
#> 1120     新疆 2023 农业机械购置投入   519063.09  万元  v9_t1_gztr_jxgz
#> 1121     全国 2023     农机服务收入 48290124.50  万元 v9_t1_zsr_njfwsr
#> 1122     北京 2023     农机服务收入    17040.40  万元 v9_t1_zsr_njfwsr
#> 1123     天津 2023     农机服务收入    93807.00  万元 v9_t1_zsr_njfwsr
#> 1124     河北 2023     农机服务收入  1964047.20  万元 v9_t1_zsr_njfwsr
#> 1125     山西 2023     农机服务收入   870216.00  万元 v9_t1_zsr_njfwsr
#> 1126   内蒙古 2023     农机服务收入  1860960.09  万元 v9_t1_zsr_njfwsr
#> 1127     辽宁 2023     农机服务收入  1123882.90  万元 v9_t1_zsr_njfwsr
#> 1128     吉林 2023     农机服务收入  1770974.00  万元 v9_t1_zsr_njfwsr
#> 1129   黑龙江 2023     农机服务收入  1776831.00  万元 v9_t1_zsr_njfwsr
#> 1130     上海 2023     农机服务收入    35561.00  万元 v9_t1_zsr_njfwsr
#> 1131     江苏 2023     农机服务收入  4334624.45  万元 v9_t1_zsr_njfwsr
#> 1132     浙江 2023     农机服务收入   791236.13  万元 v9_t1_zsr_njfwsr
#> 1133     安徽 2023     农机服务收入  5555295.53  万元 v9_t1_zsr_njfwsr
#> 1134     福建 2023     农机服务收入   634188.50  万元 v9_t1_zsr_njfwsr
#> 1135     江西 2023     农机服务收入  1791193.00  万元 v9_t1_zsr_njfwsr
#> 1136     山东 2023     农机服务收入  4392255.67  万元 v9_t1_zsr_njfwsr
#> 1137     河南 2023     农机服务收入  2807993.00  万元 v9_t1_zsr_njfwsr
#> 1138     湖北 2023     农机服务收入  2436913.00  万元 v9_t1_zsr_njfwsr
#> 1139     湖南 2023     农机服务收入  3649823.00  万元 v9_t1_zsr_njfwsr
#> 1140     广东 2023     农机服务收入  1437256.59  万元 v9_t1_zsr_njfwsr
#> 1141     广西 2023     农机服务收入  2948979.05  万元 v9_t1_zsr_njfwsr
#> 1142     海南 2023     农机服务收入   332462.00  万元 v9_t1_zsr_njfwsr
#> 1143     重庆 2023     农机服务收入   379082.00  万元 v9_t1_zsr_njfwsr
#> 1144     四川 2023     农机服务收入  1756905.00  万元 v9_t1_zsr_njfwsr
#> 1145     贵州 2023     农机服务收入   270812.88  万元 v9_t1_zsr_njfwsr
#> 1146     云南 2023     农机服务收入   940986.91  万元 v9_t1_zsr_njfwsr
#> 1147     西藏 2023     农机服务收入    15277.02  万元 v9_t1_zsr_njfwsr
#> 1148     陕西 2023     农机服务收入  1072764.41  万元 v9_t1_zsr_njfwsr
#> 1149     甘肃 2023     农机服务收入   946504.00  万元 v9_t1_zsr_njfwsr
#> 1150     青海 2023     农机服务收入    85669.00  万元 v9_t1_zsr_njfwsr
#> 1151     宁夏 2023     农机服务收入   261091.00  万元 v9_t1_zsr_njfwsr
#> 1152     新疆 2023     农机服务收入  1559165.77  万元 v9_t1_zsr_njfwsr