Details of Approved List of National Agricultural Sci-tech Park

A data set containing Approved List of National Agricultural Sci-tech Park year by year from the public site https://www.most.gov.cn, with wide data format.

Usage

PubAgriParkList

格式

A data frame:

index

integer, the ordered index of list

batch

character, approved batch Number with digits 2

name

character, name of the agri park

province

character, province of agri park in reduced chinese

参考文献出处

https://www.huhuaping.com/