Agricultural Mechanization Services by Region

This dataset contains comprehensive statistics on agricultural mechanization services by region, including service organizations, personnel, investment, and output. The data is extracted from the China Agricultural Machinery Industry Yearbook, covering all provinces and regions in China.

Usage

MachineService

Format

A data frame:

province

character. Province name, including national total.

year

character. Year of the statistics.

chn_block4

character. Service indicator in Chinese (e.g., "Year-end Organizations", "Year-end Personnel", "Service Income").

value

numeric. The statistical value.

units

character. Units of measurement (e.g., "organizations", "persons", "10,000 yuan").

variables

character. Variable name in coded format.

Source

China Agricultural Machinery Industry Yearbook, China Agricultural Machinery Industry Association

Details

The dataset covers agricultural mechanization services from 2010 to the latest available year.

Service indicators include organization counts, personnel numbers, investment amounts, and service income.

Data structure evolved over time: separate tables from 2010-2013, merged tables from 2014-2017, and fully integrated tables from 2018-2023.

Data is in long format for easy analysis and visualization.

Examples

# View the structure of the dataset
str(MachineService)
#> 'data.frame':	8124 obs. of  6 variables:
#>  $ province  : chr  "全国" "北京" "天津" "河北" ...
#>  $ year      : chr  "2023" "2023" "2023" "2023" ...
#>  $ chn_block4: chr  "年末机构数A" "年末机构数A" "年末机构数A" "年末机构数A" ...
#>  $ value     : num  195606 225 214 5533 4625 ...
#>  $ units     : chr  "个" "个" "个" "个" ...
#>  $ variables : chr  "v9_t1_fwzz_jgs" "v9_t1_fwzz_jgs" "v9_t1_fwzz_jgs" "v9_t1_fwzz_jgs" ...
# Filter data for year-end organizations in 2023
MachineService[MachineService$year == "2023", ]
#>     province year       chn_block4       value units        variables
#> 1       全国 2023      年末机构数A   195606.00    个   v9_t1_fwzz_jgs
#> 2       北京 2023      年末机构数A      225.00    个   v9_t1_fwzz_jgs
#> 3       天津 2023      年末机构数A      214.00    个   v9_t1_fwzz_jgs
#> 4       河北 2023      年末机构数A     5533.00    个   v9_t1_fwzz_jgs
#> 5       山西 2023      年末机构数A     4625.00    个   v9_t1_fwzz_jgs
#> 6     内蒙古 2023      年末机构数A     4448.00    个   v9_t1_fwzz_jgs
#> 7       辽宁 2023      年末机构数A     5141.00    个   v9_t1_fwzz_jgs
#> 8       吉林 2023      年末机构数A     9731.00    个   v9_t1_fwzz_jgs
#> 9     黑龙江 2023      年末机构数A    22623.00    个   v9_t1_fwzz_jgs
#> 10      上海 2023      年末机构数A      309.00    个   v9_t1_fwzz_jgs
#> 11      江苏 2023      年末机构数A    12203.00    个   v9_t1_fwzz_jgs
#> 12      浙江 2023      年末机构数A     3519.00    个   v9_t1_fwzz_jgs
#> 13      安徽 2023      年末机构数A    11107.00    个   v9_t1_fwzz_jgs
#> 14      福建 2023      年末机构数A     1036.00    个   v9_t1_fwzz_jgs
#> 15      江西 2023      年末机构数A    12700.00    个   v9_t1_fwzz_jgs
#> 16      山东 2023      年末机构数A    21428.00    个   v9_t1_fwzz_jgs
#> 17      河南 2023      年末机构数A    13757.00    个   v9_t1_fwzz_jgs
#> 18      湖北 2023      年末机构数A     6744.00    个   v9_t1_fwzz_jgs
#> 19      湖南 2023      年末机构数A    19194.00    个   v9_t1_fwzz_jgs
#> 20      广东 2023      年末机构数A     2542.00    个   v9_t1_fwzz_jgs
#> 21      广西 2023      年末机构数A     3073.00    个   v9_t1_fwzz_jgs
#> 22      海南 2023      年末机构数A      181.00    个   v9_t1_fwzz_jgs
#> 23      重庆 2023      年末机构数A     4831.00    个   v9_t1_fwzz_jgs
#> 24      四川 2023      年末机构数A    16449.00    个   v9_t1_fwzz_jgs
#> 25      贵州 2023      年末机构数A     1951.00    个   v9_t1_fwzz_jgs
#> 26      云南 2023      年末机构数A     1000.00    个   v9_t1_fwzz_jgs
#> 27      西藏 2023      年末机构数A      231.00    个   v9_t1_fwzz_jgs
#> 28      陕西 2023      年末机构数A     2358.00    个   v9_t1_fwzz_jgs
#> 29      甘肃 2023      年末机构数A     4674.00    个   v9_t1_fwzz_jgs
#> 30      青海 2023      年末机构数A      853.00    个   v9_t1_fwzz_jgs
#> 31      宁夏 2023      年末机构数A      473.00    个   v9_t1_fwzz_jgs
#> 32      新疆 2023      年末机构数A     1676.00    个   v9_t1_fwzz_jgs
#> 33      全国 2023        年末人数A  2103348.00    人    v9_t1_fwzz_rs
#> 34      北京 2023        年末人数A     1541.00    人    v9_t1_fwzz_rs
#> 35      天津 2023        年末人数A     5456.00    人    v9_t1_fwzz_rs
#> 36      河北 2023        年末人数A    62072.00    人    v9_t1_fwzz_rs
#> 37      山西 2023        年末人数A    30531.00    人    v9_t1_fwzz_rs
#> 38    内蒙古 2023        年末人数A    44837.00    人    v9_t1_fwzz_rs
#> 39      辽宁 2023        年末人数A    59506.00    人    v9_t1_fwzz_rs
#> 40      吉林 2023        年末人数A    94947.00    人    v9_t1_fwzz_rs
#> 41    黑龙江 2023        年末人数A    92848.00    人    v9_t1_fwzz_rs
#> 42      上海 2023        年末人数A     2186.00    人    v9_t1_fwzz_rs
#> 43      江苏 2023        年末人数A   410864.00    人    v9_t1_fwzz_rs
#> 44      浙江 2023        年末人数A    28043.00    人    v9_t1_fwzz_rs
#> 45      安徽 2023        年末人数A   125061.00    人    v9_t1_fwzz_rs
#> 46      福建 2023        年末人数A    23205.00    人    v9_t1_fwzz_rs
#> 47      江西 2023        年末人数A    82445.00    人    v9_t1_fwzz_rs
#> 48      山东 2023        年末人数A   209603.00    人    v9_t1_fwzz_rs
#> 49      河南 2023        年末人数A   154356.00    人    v9_t1_fwzz_rs
#> 50      湖北 2023        年末人数A   148832.00    人    v9_t1_fwzz_rs
#> 51      湖南 2023        年末人数A   132682.00    人    v9_t1_fwzz_rs
#> 52      广东 2023        年末人数A    27958.00    人    v9_t1_fwzz_rs
#> 53      广西 2023        年末人数A    42799.00    人    v9_t1_fwzz_rs
#> 54      海南 2023        年末人数A     1158.00    人    v9_t1_fwzz_rs
#> 55      重庆 2023        年末人数A    86494.00    人    v9_t1_fwzz_rs
#> 56      四川 2023        年末人数A    84735.00    人    v9_t1_fwzz_rs
#> 57      贵州 2023        年末人数A    21016.00    人    v9_t1_fwzz_rs
#> 58      云南 2023        年末人数A    14836.00    人    v9_t1_fwzz_rs
#> 59      西藏 2023        年末人数A     2467.00    人    v9_t1_fwzz_rs
#> 60      陕西 2023        年末人数A    36668.00    人    v9_t1_fwzz_rs
#> 61      甘肃 2023        年末人数A    38792.00    人    v9_t1_fwzz_rs
#> 62      青海 2023        年末人数A     4976.00    人    v9_t1_fwzz_rs
#> 63      宁夏 2023        年末人数A     6700.00    人    v9_t1_fwzz_rs
#> 64      新疆 2023        年末人数A    21102.00    人    v9_t1_fwzz_rs
#> 65      全国 2023      年末机构数B    79470.00    个  v9_t1_zyhzs_jgs
#> 66      北京 2023      年末机构数B      166.00    个  v9_t1_zyhzs_jgs
#> 67      天津 2023      年末机构数B      163.00    个  v9_t1_zyhzs_jgs
#> 68      河北 2023      年末机构数B     2624.00    个  v9_t1_zyhzs_jgs
#> 69      山西 2023      年末机构数B     2777.00    个  v9_t1_zyhzs_jgs
#> 70    内蒙古 2023      年末机构数B     3220.00    个  v9_t1_zyhzs_jgs
#> 71      辽宁 2023      年末机构数B     3914.00    个  v9_t1_zyhzs_jgs
#> 72      吉林 2023      年末机构数B     6721.00    个  v9_t1_zyhzs_jgs
#> 73    黑龙江 2023      年末机构数B     2299.00    个  v9_t1_zyhzs_jgs
#> 74      上海 2023      年末机构数B      189.00    个  v9_t1_zyhzs_jgs
#> 75      江苏 2023      年末机构数B     6925.00    个  v9_t1_zyhzs_jgs
#> 76      浙江 2023      年末机构数B     1454.00    个  v9_t1_zyhzs_jgs
#> 77      安徽 2023      年末机构数B     6201.00    个  v9_t1_zyhzs_jgs
#> 78      福建 2023      年末机构数B      760.00    个  v9_t1_zyhzs_jgs
#> 79      江西 2023      年末机构数B     1663.00    个  v9_t1_zyhzs_jgs
#> 80      山东 2023      年末机构数B     9217.00    个  v9_t1_zyhzs_jgs
#> 81      河南 2023      年末机构数B     7640.00    个  v9_t1_zyhzs_jgs
#> 82      湖北 2023      年末机构数B     3588.00    个  v9_t1_zyhzs_jgs
#> 83      湖南 2023      年末机构数B     5780.00    个  v9_t1_zyhzs_jgs
#> 84      广东 2023      年末机构数B     1071.00    个  v9_t1_zyhzs_jgs
#> 85      广西 2023      年末机构数B     1691.00    个  v9_t1_zyhzs_jgs
#> 86      海南 2023      年末机构数B       88.00    个  v9_t1_zyhzs_jgs
#> 87      重庆 2023      年末机构数B     1053.00    个  v9_t1_zyhzs_jgs
#> 88      四川 2023      年末机构数B     1763.00    个  v9_t1_zyhzs_jgs
#> 89      贵州 2023      年末机构数B     1008.00    个  v9_t1_zyhzs_jgs
#> 90      云南 2023      年末机构数B      551.00    个  v9_t1_zyhzs_jgs
#> 91      西藏 2023      年末机构数B      169.00    个  v9_t1_zyhzs_jgs
#> 92      陕西 2023      年末机构数B     1487.00    个  v9_t1_zyhzs_jgs
#> 93      甘肃 2023      年末机构数B     3057.00    个  v9_t1_zyhzs_jgs
#> 94      青海 2023      年末机构数B      195.00    个  v9_t1_zyhzs_jgs
#> 95      宁夏 2023      年末机构数B      275.00    个  v9_t1_zyhzs_jgs
#> 96      新疆 2023      年末机构数B     1440.00    个  v9_t1_zyhzs_jgs
#> 97      全国 2023        年末人数B  1412352.00    人   v9_t1_zyhzs_rs
#> 98      北京 2023        年末人数B      928.00    人   v9_t1_zyhzs_rs
#> 99      天津 2023        年末人数B     4893.00    人   v9_t1_zyhzs_rs
#> 100     河北 2023        年末人数B    46925.00    人   v9_t1_zyhzs_rs
#> 101     山西 2023        年末人数B    24179.00    人   v9_t1_zyhzs_rs
#> 102   内蒙古 2023        年末人数B    34294.00    人   v9_t1_zyhzs_rs
#> 103     辽宁 2023        年末人数B    48257.00    人   v9_t1_zyhzs_rs
#> 104     吉林 2023        年末人数B    68445.00    人   v9_t1_zyhzs_rs
#> 105   黑龙江 2023        年末人数B    24231.00    人   v9_t1_zyhzs_rs
#> 106     上海 2023        年末人数B     1477.00    人   v9_t1_zyhzs_rs
#> 107     江苏 2023        年末人数B   297051.00    人   v9_t1_zyhzs_rs
#> 108     浙江 2023        年末人数B    18312.00    人   v9_t1_zyhzs_rs
#> 109     安徽 2023        年末人数B    87028.00    人   v9_t1_zyhzs_rs
#> 110     福建 2023        年末人数B    20002.00    人   v9_t1_zyhzs_rs
#> 111     江西 2023        年末人数B    27631.00    人   v9_t1_zyhzs_rs
#> 112     山东 2023        年末人数B   143471.00    人   v9_t1_zyhzs_rs
#> 113     河南 2023        年末人数B   125111.00    人   v9_t1_zyhzs_rs
#> 114     湖北 2023        年末人数B   110230.00    人   v9_t1_zyhzs_rs
#> 115     湖南 2023        年末人数B    77240.00    人   v9_t1_zyhzs_rs
#> 116     广东 2023        年末人数B    19508.00    人   v9_t1_zyhzs_rs
#> 117     广西 2023        年末人数B    25100.00    人   v9_t1_zyhzs_rs
#> 118     海南 2023        年末人数B      798.00    人   v9_t1_zyhzs_rs
#> 119     重庆 2023        年末人数B    56949.00    人   v9_t1_zyhzs_rs
#> 120     四川 2023        年末人数B    34693.00    人   v9_t1_zyhzs_rs
#> 121     贵州 2023        年末人数B    10533.00    人   v9_t1_zyhzs_rs
#> 122     云南 2023        年末人数B    11801.00    人   v9_t1_zyhzs_rs
#> 123     西藏 2023        年末人数B     2301.00    人   v9_t1_zyhzs_rs
#> 124     陕西 2023        年末人数B    29798.00    人   v9_t1_zyhzs_rs
#> 125     甘肃 2023        年末人数B    30732.00    人   v9_t1_zyhzs_rs
#> 126     青海 2023        年末人数B     3069.00    人   v9_t1_zyhzs_rs
#> 127     宁夏 2023        年末人数B     3933.00    人   v9_t1_zyhzs_rs
#> 128     新疆 2023        年末人数B    19113.00    人   v9_t1_zyhzs_rs
#> 129     全国 2023      年末机构数C 39193020.00    个    v9_t1_njh_jgs
#> 130     北京 2023      年末机构数C    11788.00    个    v9_t1_njh_jgs
#> 131     天津 2023      年末机构数C    34668.00    个    v9_t1_njh_jgs
#> 132     河北 2023      年末机构数C  2420982.00    个    v9_t1_njh_jgs
#> 133     山西 2023      年末机构数C   510702.00    个    v9_t1_njh_jgs
#> 134   内蒙古 2023      年末机构数C  1248251.00    个    v9_t1_njh_jgs
#> 135     辽宁 2023      年末机构数C   536991.00    个    v9_t1_njh_jgs
#> 136     吉林 2023      年末机构数C  1100789.00    个    v9_t1_njh_jgs
#> 137   黑龙江 2023      年末机构数C  1169734.00    个    v9_t1_njh_jgs
#> 138     上海 2023      年末机构数C     4462.00    个    v9_t1_njh_jgs
#> 139     江苏 2023      年末机构数C  1038229.00    个    v9_t1_njh_jgs
#> 140     浙江 2023      年末机构数C   466712.00    个    v9_t1_njh_jgs
#> 141     安徽 2023      年末机构数C  3117163.00    个    v9_t1_njh_jgs
#> 142     福建 2023      年末机构数C   600506.00    个    v9_t1_njh_jgs
#> 143     江西 2023      年末机构数C  1023125.00    个    v9_t1_njh_jgs
#> 144     山东 2023      年末机构数C  3914821.00    个    v9_t1_njh_jgs
#> 145     河南 2023      年末机构数C  4890000.00    个    v9_t1_njh_jgs
#> 146     湖北 2023      年末机构数C  1724808.00    个    v9_t1_njh_jgs
#> 147     湖南 2023      年末机构数C  2046751.00    个    v9_t1_njh_jgs
#> 148     广东 2023      年末机构数C  1042910.00    个    v9_t1_njh_jgs
#> 149     广西 2023      年末机构数C  2468660.00    个    v9_t1_njh_jgs
#> 150     海南 2023      年末机构数C   208263.00    个    v9_t1_njh_jgs
#> 151     重庆 2023      年末机构数C  1212476.00    个    v9_t1_njh_jgs
#> 152     四川 2023      年末机构数C  2394351.00    个    v9_t1_njh_jgs
#> 153     贵州 2023      年末机构数C   572674.00    个    v9_t1_njh_jgs
#> 154     云南 2023      年末机构数C  1969794.00    个    v9_t1_njh_jgs
#> 155     西藏 2023      年末机构数C   228953.00    个    v9_t1_njh_jgs
#> 156     陕西 2023      年末机构数C   956154.00    个    v9_t1_njh_jgs
#> 157     甘肃 2023      年末机构数C  1211975.00    个    v9_t1_njh_jgs
#> 158     青海 2023      年末机构数C   196380.00    个    v9_t1_njh_jgs
#> 159     宁夏 2023      年末机构数C   252033.00    个    v9_t1_njh_jgs
#> 160     新疆 2023      年末机构数C   590416.00    个    v9_t1_njh_jgs
#> 161     全国 2023        年末人数C 46615424.00    人     v9_t1_njh_rs
#> 162     北京 2023        年末人数C    12561.00    人     v9_t1_njh_rs
#> 163     天津 2023        年末人数C    41449.00    人     v9_t1_njh_rs
#> 164     河北 2023        年末人数C  2808740.00    人     v9_t1_njh_rs
#> 165     山西 2023        年末人数C   648950.00    人     v9_t1_njh_rs
#> 166   内蒙古 2023        年末人数C  1587289.00    人     v9_t1_njh_rs
#> 167     辽宁 2023        年末人数C   651342.00    人     v9_t1_njh_rs
#> 168     吉林 2023        年末人数C  1305715.00    人     v9_t1_njh_rs
#> 169   黑龙江 2023        年末人数C  1458849.00    人     v9_t1_njh_rs
#> 170     上海 2023        年末人数C     6036.00    人     v9_t1_njh_rs
#> 171     江苏 2023        年末人数C  1242430.00    人     v9_t1_njh_rs
#> 172     浙江 2023        年末人数C   539734.00    人     v9_t1_njh_rs
#> 173     安徽 2023        年末人数C  3651851.00    人     v9_t1_njh_rs
#> 174     福建 2023        年末人数C   672279.00    人     v9_t1_njh_rs
#> 175     江西 2023        年末人数C  1401998.00    人     v9_t1_njh_rs
#> 176     山东 2023        年末人数C  4545244.00    人     v9_t1_njh_rs
#> 177     河南 2023        年末人数C  5489074.00    人     v9_t1_njh_rs
#> 178     湖北 2023        年末人数C  2239058.00    人     v9_t1_njh_rs
#> 179     湖南 2023        年末人数C  2716203.00    人     v9_t1_njh_rs
#> 180     广东 2023        年末人数C  1198857.00    人     v9_t1_njh_rs
#> 181     广西 2023        年末人数C  3153394.00    人     v9_t1_njh_rs
#> 182     海南 2023        年末人数C   239341.00    人     v9_t1_njh_rs
#> 183     重庆 2023        年末人数C  1324445.00    人     v9_t1_njh_rs
#> 184     四川 2023        年末人数C  2802935.00    人     v9_t1_njh_rs
#> 185     贵州 2023        年末人数C   818626.00    人     v9_t1_njh_rs
#> 186     云南 2023        年末人数C  2099837.00    人     v9_t1_njh_rs
#> 187     西藏 2023        年末人数C   268726.00    人     v9_t1_njh_rs
#> 188     陕西 2023        年末人数C  1100590.00    人     v9_t1_njh_rs
#> 189     甘肃 2023        年末人数C  1325238.00    人     v9_t1_njh_rs
#> 190     青海 2023        年末人数C   219044.00    人     v9_t1_njh_rs
#> 191     宁夏 2023        年末人数C   354965.00    人     v9_t1_njh_rs
#> 192     新疆 2023        年末人数C   649860.00    人     v9_t1_njh_rs
#> 193     全国 2023      年末机构数D  4086769.00    个    v9_t1_zyh_jgs
#> 194     北京 2023      年末机构数D      512.00    个    v9_t1_zyh_jgs
#> 195     天津 2023      年末机构数D     7784.00    个    v9_t1_zyh_jgs
#> 196     河北 2023      年末机构数D   179213.00    个    v9_t1_zyh_jgs
#> 197     山西 2023      年末机构数D    70004.00    个    v9_t1_zyh_jgs
#> 198   内蒙古 2023      年末机构数D   108438.00    个    v9_t1_zyh_jgs
#> 199     辽宁 2023      年末机构数D    59143.00    个    v9_t1_zyh_jgs
#> 200     吉林 2023      年末机构数D    29654.00    个    v9_t1_zyh_jgs
#> 201   黑龙江 2023      年末机构数D   135036.00    个    v9_t1_zyh_jgs
#> 202     上海 2023      年末机构数D      940.00    个    v9_t1_zyh_jgs
#> 203     江苏 2023      年末机构数D   176342.00    个    v9_t1_zyh_jgs
#> 204     浙江 2023      年末机构数D    81312.00    个    v9_t1_zyh_jgs
#> 205     安徽 2023      年末机构数D   417083.00    个    v9_t1_zyh_jgs
#> 206     福建 2023      年末机构数D    89956.00    个    v9_t1_zyh_jgs
#> 207     江西 2023      年末机构数D   330383.00    个    v9_t1_zyh_jgs
#> 208     山东 2023      年末机构数D   560044.00    个    v9_t1_zyh_jgs
#> 209     河南 2023      年末机构数D   232377.00    个    v9_t1_zyh_jgs
#> 210     湖北 2023      年末机构数D   166335.00    个    v9_t1_zyh_jgs
#> 211     湖南 2023      年末机构数D   180452.00    个    v9_t1_zyh_jgs
#> 212     广东 2023      年末机构数D    51114.00    个    v9_t1_zyh_jgs
#> 213     广西 2023      年末机构数D    29625.00    个    v9_t1_zyh_jgs
#> 214     海南 2023      年末机构数D    47100.00    个    v9_t1_zyh_jgs
#> 215     重庆 2023      年末机构数D   140610.00    个    v9_t1_zyh_jgs
#> 216     四川 2023      年末机构数D   189595.00    个    v9_t1_zyh_jgs
#> 217     贵州 2023      年末机构数D   157206.00    个    v9_t1_zyh_jgs
#> 218     云南 2023      年末机构数D    61343.00    个    v9_t1_zyh_jgs
#> 219     西藏 2023      年末机构数D      183.00    个    v9_t1_zyh_jgs
#> 220     陕西 2023      年末机构数D   222268.00    个    v9_t1_zyh_jgs
#> 221     甘肃 2023      年末机构数D   191079.00    个    v9_t1_zyh_jgs
#> 222     青海 2023      年末机构数D     4286.00    个    v9_t1_zyh_jgs
#> 223     宁夏 2023      年末机构数D    37906.00    个    v9_t1_zyh_jgs
#> 224     新疆 2023      年末机构数D   121330.00    个    v9_t1_zyh_jgs
#> 225     全国 2023        年末人数D  5732807.00    人     v9_t1_zyh_rs
#> 226     北京 2023        年末人数D      769.00    人     v9_t1_zyh_rs
#> 227     天津 2023        年末人数D    11129.00    人     v9_t1_zyh_rs
#> 228     河北 2023        年末人数D   277096.00    人     v9_t1_zyh_rs
#> 229     山西 2023        年末人数D    94591.00    人     v9_t1_zyh_rs
#> 230   内蒙古 2023        年末人数D   155893.00    人     v9_t1_zyh_rs
#> 231     辽宁 2023        年末人数D    86064.00    人     v9_t1_zyh_rs
#> 232     吉林 2023        年末人数D    32335.00    人     v9_t1_zyh_rs
#> 233   黑龙江 2023        年末人数D   195571.00    人     v9_t1_zyh_rs
#> 234     上海 2023        年末人数D     1428.00    人     v9_t1_zyh_rs
#> 235     江苏 2023        年末人数D   264692.00    人     v9_t1_zyh_rs
#> 236     浙江 2023        年末人数D    95903.00    人     v9_t1_zyh_rs
#> 237     安徽 2023        年末人数D   586861.00    人     v9_t1_zyh_rs
#> 238     福建 2023        年末人数D   110358.00    人     v9_t1_zyh_rs
#> 239     江西 2023        年末人数D   495339.00    人     v9_t1_zyh_rs
#> 240     山东 2023        年末人数D   782028.00    人     v9_t1_zyh_rs
#> 241     河南 2023        年末人数D   372472.00    人     v9_t1_zyh_rs
#> 242     湖北 2023        年末人数D   325121.00    人     v9_t1_zyh_rs
#> 243     湖南 2023        年末人数D   270372.00    人     v9_t1_zyh_rs
#> 244     广东 2023        年末人数D    82531.00    人     v9_t1_zyh_rs
#> 245     广西 2023        年末人数D    68852.00    人     v9_t1_zyh_rs
#> 246     海南 2023        年末人数D    51173.00    人     v9_t1_zyh_rs
#> 247     重庆 2023        年末人数D   167567.00    人     v9_t1_zyh_rs
#> 248     四川 2023        年末人数D   242128.00    人     v9_t1_zyh_rs
#> 249     贵州 2023        年末人数D   209929.00    人     v9_t1_zyh_rs
#> 250     云南 2023        年末人数D    70021.00    人     v9_t1_zyh_rs
#> 251     西藏 2023        年末人数D      268.00    人     v9_t1_zyh_rs
#> 252     陕西 2023        年末人数D   244540.00    人     v9_t1_zyh_rs
#> 253     甘肃 2023        年末人数D   221861.00    人     v9_t1_zyh_rs
#> 254     青海 2023        年末人数D     5937.00    人     v9_t1_zyh_rs
#> 255     宁夏 2023        年末人数D    41913.00    人     v9_t1_zyh_rs
#> 256     新疆 2023        年末人数D   155313.00    人     v9_t1_zyh_rs
#> 257     全国 2023        年末人数G 49928434.00    人  v9_t1_njcyry_rs
#> 258     北京 2023        年末人数G    17604.00    人  v9_t1_njcyry_rs
#> 259     天津 2023        年末人数G    60986.00    人  v9_t1_njcyry_rs
#> 260     河北 2023        年末人数G  3268079.00    人  v9_t1_njcyry_rs
#> 261     山西 2023        年末人数G   704283.00    人  v9_t1_njcyry_rs
#> 262   内蒙古 2023        年末人数G  1797205.00    人  v9_t1_njcyry_rs
#> 263     辽宁 2023        年末人数G   831064.00    人  v9_t1_njcyry_rs
#> 264     吉林 2023        年末人数G  1353191.00    人  v9_t1_njcyry_rs
#> 265   黑龙江 2023        年末人数G  1741918.00    人  v9_t1_njcyry_rs
#> 266     上海 2023        年末人数G    14136.00    人  v9_t1_njcyry_rs
#> 267     江苏 2023        年末人数G  1559645.00    人  v9_t1_njcyry_rs
#> 268     浙江 2023        年末人数G   546315.00    人  v9_t1_njcyry_rs
#> 269     安徽 2023        年末人数G  3840352.00    人  v9_t1_njcyry_rs
#> 270     福建 2023        年末人数G   698424.00    人  v9_t1_njcyry_rs
#> 271     江西 2023        年末人数G  1403260.00    人  v9_t1_njcyry_rs
#> 272     山东 2023        年末人数G  4675254.00    人  v9_t1_njcyry_rs
#> 273     河南 2023        年末人数G  5774998.00    人  v9_t1_njcyry_rs
#> 274     湖北 2023        年末人数G  2249249.00    人  v9_t1_njcyry_rs
#> 275     湖南 2023        年末人数G  2799088.00    人  v9_t1_njcyry_rs
#> 276     广东 2023        年末人数G  1226209.00    人  v9_t1_njcyry_rs
#> 277     广西 2023        年末人数G  3295941.00    人  v9_t1_njcyry_rs
#> 278     海南 2023        年末人数G   243491.00    人  v9_t1_njcyry_rs
#> 279     重庆 2023        年末人数G  1377979.00    人  v9_t1_njcyry_rs
#> 280     四川 2023        年末人数G  2841383.00    人  v9_t1_njcyry_rs
#> 281     贵州 2023        年末人数G  1114734.00    人  v9_t1_njcyry_rs
#> 282     云南 2023        年末人数G  2251208.00    人  v9_t1_njcyry_rs
#> 283     西藏 2023        年末人数G   321149.00    人  v9_t1_njcyry_rs
#> 284     陕西 2023        年末人数G  1104253.00    人  v9_t1_njcyry_rs
#> 285     甘肃 2023        年末人数G  1393248.00    人  v9_t1_njcyry_rs
#> 286     青海 2023        年末人数G   268423.00    人  v9_t1_njcyry_rs
#> 287     宁夏 2023        年末人数G   359575.00    人  v9_t1_njcyry_rs
#> 288     新疆 2023        年末人数G   690707.00    人  v9_t1_njcyry_rs
#> 289     全国 2023      年末机构数F   142782.00    个   v9_t1_njwx_jgs
#> 290     北京 2023      年末机构数F      138.00    个   v9_t1_njwx_jgs
#> 291     天津 2023      年末机构数F      183.00    个   v9_t1_njwx_jgs
#> 292     河北 2023      年末机构数F    11713.00    个   v9_t1_njwx_jgs
#> 293     山西 2023      年末机构数F     4809.00    个   v9_t1_njwx_jgs
#> 294   内蒙古 2023      年末机构数F     6481.00    个   v9_t1_njwx_jgs
#> 295     辽宁 2023      年末机构数F     3158.00    个   v9_t1_njwx_jgs
#> 296     吉林 2023      年末机构数F     6342.00    个   v9_t1_njwx_jgs
#> 297   黑龙江 2023      年末机构数F     6545.00    个   v9_t1_njwx_jgs
#> 298     上海 2023      年末机构数F       25.00    个   v9_t1_njwx_jgs
#> 299     江苏 2023      年末机构数F     3145.00    个   v9_t1_njwx_jgs
#> 300     浙江 2023      年末机构数F     1830.00    个   v9_t1_njwx_jgs
#> 301     安徽 2023      年末机构数F     9116.00    个   v9_t1_njwx_jgs
#> 302     福建 2023      年末机构数F      291.00    个   v9_t1_njwx_jgs
#> 303     江西 2023      年末机构数F     9036.00    个   v9_t1_njwx_jgs
#> 304     山东 2023      年末机构数F     4031.00    个   v9_t1_njwx_jgs
#> 305     河南 2023      年末机构数F    18279.00    个   v9_t1_njwx_jgs
#> 306     湖北 2023      年末机构数F     5139.00    个   v9_t1_njwx_jgs
#> 307     湖南 2023      年末机构数F     5045.00    个   v9_t1_njwx_jgs
#> 308     广东 2023      年末机构数F     6576.00    个   v9_t1_njwx_jgs
#> 309     广西 2023      年末机构数F     3587.00    个   v9_t1_njwx_jgs
#> 310     海南 2023      年末机构数F      736.00    个   v9_t1_njwx_jgs
#> 311     重庆 2023      年末机构数F     2529.00    个   v9_t1_njwx_jgs
#> 312     四川 2023      年末机构数F     7531.00    个   v9_t1_njwx_jgs
#> 313     贵州 2023      年末机构数F     4342.00    个   v9_t1_njwx_jgs
#> 314     云南 2023      年末机构数F     8535.00    个   v9_t1_njwx_jgs
#> 315     西藏 2023      年末机构数F      234.00    个   v9_t1_njwx_jgs
#> 316     陕西 2023      年末机构数F     3494.00    个   v9_t1_njwx_jgs
#> 317     甘肃 2023      年末机构数F     4722.00    个   v9_t1_njwx_jgs
#> 318     青海 2023      年末机构数F     1042.00    个   v9_t1_njwx_jgs
#> 319     宁夏 2023      年末机构数F     1713.00    个   v9_t1_njwx_jgs
#> 320     新疆 2023      年末机构数F     2057.00    个   v9_t1_njwx_jgs
#> 321     全国 2023        年末人数F   355285.00    人    v9_t1_njwx_rs
#> 322     北京 2023        年末人数F      278.00    人    v9_t1_njwx_rs
#> 323     天津 2023        年末人数F      387.00    人    v9_t1_njwx_rs
#> 324     河北 2023        年末人数F    28873.00    人    v9_t1_njwx_rs
#> 325     山西 2023        年末人数F    12012.00    人    v9_t1_njwx_rs
#> 326   内蒙古 2023        年末人数F    13664.00    人    v9_t1_njwx_rs
#> 327     辽宁 2023        年末人数F     7272.00    人    v9_t1_njwx_rs
#> 328     吉林 2023        年末人数F    14669.00    人    v9_t1_njwx_rs
#> 329   黑龙江 2023        年末人数F    20638.00    人    v9_t1_njwx_rs
#> 330     上海 2023        年末人数F       93.00    人    v9_t1_njwx_rs
#> 331     江苏 2023        年末人数F    10279.00    人    v9_t1_njwx_rs
#> 332     浙江 2023        年末人数F     3376.00    人    v9_t1_njwx_rs
#> 333     安徽 2023        年末人数F    21792.00    人    v9_t1_njwx_rs
#> 334     福建 2023        年末人数F      531.00    人    v9_t1_njwx_rs
#> 335     江西 2023        年末人数F    28113.00    人    v9_t1_njwx_rs
#> 336     山东 2023        年末人数F    11936.00    人    v9_t1_njwx_rs
#> 337     河南 2023        年末人数F    38233.00    人    v9_t1_njwx_rs
#> 338     湖北 2023        年末人数F    15925.00    人    v9_t1_njwx_rs
#> 339     湖南 2023        年末人数F    13864.00    人    v9_t1_njwx_rs
#> 340     广东 2023        年末人数F    16901.00    人    v9_t1_njwx_rs
#> 341     广西 2023        年末人数F     8789.00    人    v9_t1_njwx_rs
#> 342     海南 2023        年末人数F     1818.00    人    v9_t1_njwx_rs
#> 343     重庆 2023        年末人数F     7229.00    人    v9_t1_njwx_rs
#> 344     四川 2023        年末人数F    22965.00    人    v9_t1_njwx_rs
#> 345     贵州 2023        年末人数F    11221.00    人    v9_t1_njwx_rs
#> 346     云南 2023        年末人数F    18536.00    人    v9_t1_njwx_rs
#> 347     西藏 2023        年末人数F      722.00    人    v9_t1_njwx_rs
#> 348     陕西 2023        年末人数F     7252.00    人    v9_t1_njwx_rs
#> 349     甘肃 2023        年末人数F     8135.00    人    v9_t1_njwx_rs
#> 350     青海 2023        年末人数F     1638.00    人    v9_t1_njwx_rs
#> 351     宁夏 2023        年末人数F     3147.00    人    v9_t1_njwx_rs
#> 352     新疆 2023        年末人数F     4043.00    人    v9_t1_njwx_rs
#> 353     全国 2023     财政资金投入   933466.46  万元  v9_t1_cztr_czzj
#> 354     北京 2023     财政资金投入     5836.71  万元  v9_t1_cztr_czzj
#> 355     天津 2023     财政资金投入     5177.81  万元  v9_t1_cztr_czzj
#> 356     河北 2023     财政资金投入    14183.60  万元  v9_t1_cztr_czzj
#> 357     山西 2023     财政资金投入    75801.00  万元  v9_t1_cztr_czzj
#> 358   内蒙古 2023     财政资金投入   105765.96  万元  v9_t1_cztr_czzj
#> 359     辽宁 2023     财政资金投入    49133.76  万元  v9_t1_cztr_czzj
#> 360     吉林 2023     财政资金投入    70608.16  万元  v9_t1_cztr_czzj
#> 361   黑龙江 2023     财政资金投入    11717.38  万元  v9_t1_cztr_czzj
#> 362     上海 2023     财政资金投入    10236.00  万元  v9_t1_cztr_czzj
#> 363     江苏 2023     财政资金投入   163624.23  万元  v9_t1_cztr_czzj
#> 364     浙江 2023     财政资金投入    16941.13  万元  v9_t1_cztr_czzj
#> 365     安徽 2023     财政资金投入    23711.57  万元  v9_t1_cztr_czzj
#> 366     福建 2023     财政资金投入     3797.57  万元  v9_t1_cztr_czzj
#> 367     江西 2023     财政资金投入    17675.00  万元  v9_t1_cztr_czzj
#> 368     山东 2023     财政资金投入    27693.65  万元  v9_t1_cztr_czzj
#> 369     河南 2023     财政资金投入    42346.90  万元  v9_t1_cztr_czzj
#> 370     湖北 2023     财政资金投入    59042.00  万元  v9_t1_cztr_czzj
#> 371     湖南 2023     财政资金投入    62593.24  万元  v9_t1_cztr_czzj
#> 372     广东 2023     财政资金投入     9247.99  万元  v9_t1_cztr_czzj
#> 373     广西 2023     财政资金投入     6133.49  万元  v9_t1_cztr_czzj
#> 374     海南 2023     财政资金投入     3965.21  万元  v9_t1_cztr_czzj
#> 375     重庆 2023     财政资金投入     9456.02  万元  v9_t1_cztr_czzj
#> 376     四川 2023     财政资金投入    36042.00  万元  v9_t1_cztr_czzj
#> 377     贵州 2023     财政资金投入    19299.73  万元  v9_t1_cztr_czzj
#> 378     云南 2023     财政资金投入     6766.75  万元  v9_t1_cztr_czzj
#> 379     西藏 2023     财政资金投入        0.01  万元  v9_t1_cztr_czzj
#> 380     陕西 2023     财政资金投入    27212.09  万元  v9_t1_cztr_czzj
#> 381     甘肃 2023     财政资金投入    19644.00  万元  v9_t1_cztr_czzj
#> 382     青海 2023     财政资金投入     3582.10  万元  v9_t1_cztr_czzj
#> 383     宁夏 2023     财政资金投入    10793.00  万元  v9_t1_cztr_czzj
#> 384     新疆 2023     财政资金投入    12962.91  万元  v9_t1_cztr_czzj
#> 385     全国 2023         科研投入    63553.55  万元  v9_t1_cztr_kytr
#> 386     北京 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 387     天津 2023         科研投入       82.81  万元  v9_t1_cztr_kytr
#> 388     河北 2023         科研投入      592.30  万元  v9_t1_cztr_kytr
#> 389     山西 2023         科研投入      485.00  万元  v9_t1_cztr_kytr
#> 390   内蒙古 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 391     辽宁 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 392     吉林 2023         科研投入      595.57  万元  v9_t1_cztr_kytr
#> 393   黑龙江 2023         科研投入     3050.00  万元  v9_t1_cztr_kytr
#> 394     上海 2023         科研投入      744.00  万元  v9_t1_cztr_kytr
#> 395     江苏 2023         科研投入     6143.00  万元  v9_t1_cztr_kytr
#> 396     浙江 2023         科研投入     1812.60  万元  v9_t1_cztr_kytr
#> 397     安徽 2023         科研投入     1904.60  万元  v9_t1_cztr_kytr
#> 398     福建 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 399     江西 2023         科研投入     2486.00  万元  v9_t1_cztr_kytr
#> 400     山东 2023         科研投入    10953.77  万元  v9_t1_cztr_kytr
#> 401     河南 2023         科研投入     2090.00  万元  v9_t1_cztr_kytr
#> 402     湖北 2023         科研投入     2783.00  万元  v9_t1_cztr_kytr
#> 403     湖南 2023         科研投入    11295.00  万元  v9_t1_cztr_kytr
#> 404     广东 2023         科研投入     3469.30  万元  v9_t1_cztr_kytr
#> 405     广西 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 406     海南 2023         科研投入     1698.00  万元  v9_t1_cztr_kytr
#> 407     重庆 2023         科研投入     1680.00  万元  v9_t1_cztr_kytr
#> 408     四川 2023         科研投入     2111.10  万元  v9_t1_cztr_kytr
#> 409     贵州 2023         科研投入      296.00  万元  v9_t1_cztr_kytr
#> 410     云南 2023         科研投入       20.00  万元  v9_t1_cztr_kytr
#> 411     西藏 2023         科研投入        0.01  万元  v9_t1_cztr_kytr
#> 412     陕西 2023         科研投入      152.00  万元  v9_t1_cztr_kytr
#> 413     甘肃 2023         科研投入     1157.00  万元  v9_t1_cztr_kytr
#> 414     青海 2023         科研投入       30.00  万元  v9_t1_cztr_kytr
#> 415     宁夏 2023         科研投入       34.00  万元  v9_t1_cztr_kytr
#> 416     新疆 2023         科研投入     6880.50  万元  v9_t1_cztr_kytr
#> 417     全国 2023         推广投入   481547.21  万元  v9_t1_cztr_tgtr
#> 418     北京 2023         推广投入     5768.81  万元  v9_t1_cztr_tgtr
#> 419     天津 2023         推广投入       88.30  万元  v9_t1_cztr_tgtr
#> 420     河北 2023         推广投入     4920.15  万元  v9_t1_cztr_tgtr
#> 421     山西 2023         推广投入     9585.00  万元  v9_t1_cztr_tgtr
#> 422   内蒙古 2023         推广投入   104431.50  万元  v9_t1_cztr_tgtr
#> 423     辽宁 2023         推广投入    16568.57  万元  v9_t1_cztr_tgtr
#> 424     吉林 2023         推广投入    69584.17  万元  v9_t1_cztr_tgtr
#> 425   黑龙江 2023         推广投入      897.46  万元  v9_t1_cztr_tgtr
#> 426     上海 2023         推广投入     8684.00  万元  v9_t1_cztr_tgtr
#> 427     江苏 2023         推广投入   101224.28  万元  v9_t1_cztr_tgtr
#> 428     浙江 2023         推广投入    14493.08  万元  v9_t1_cztr_tgtr
#> 429     安徽 2023         推广投入    10175.65  万元  v9_t1_cztr_tgtr
#> 430     福建 2023         推广投入     2856.82  万元  v9_t1_cztr_tgtr
#> 431     江西 2023         推广投入     9940.00  万元  v9_t1_cztr_tgtr
#> 432     山东 2023         推广投入     9815.14  万元  v9_t1_cztr_tgtr
#> 433     河南 2023         推广投入     2989.00  万元  v9_t1_cztr_tgtr
#> 434     湖北 2023         推广投入    11065.00  万元  v9_t1_cztr_tgtr
#> 435     湖南 2023         推广投入    15744.31  万元  v9_t1_cztr_tgtr
#> 436     广东 2023         推广投入     3651.37  万元  v9_t1_cztr_tgtr
#> 437     广西 2023         推广投入     5539.50  万元  v9_t1_cztr_tgtr
#> 438     海南 2023         推广投入      960.00  万元  v9_t1_cztr_tgtr
#> 439     重庆 2023         推广投入     4791.00  万元  v9_t1_cztr_tgtr
#> 440     四川 2023         推广投入    11544.00  万元  v9_t1_cztr_tgtr
#> 441     贵州 2023         推广投入    15228.03  万元  v9_t1_cztr_tgtr
#> 442     云南 2023         推广投入     6328.25  万元  v9_t1_cztr_tgtr
#> 443     西藏 2023         推广投入        0.01  万元  v9_t1_cztr_tgtr
#> 444     陕西 2023         推广投入    12997.06  万元  v9_t1_cztr_tgtr
#> 445     甘肃 2023         推广投入    11960.00  万元  v9_t1_cztr_tgtr
#> 446     青海 2023         推广投入     1540.10  万元  v9_t1_cztr_tgtr
#> 447     宁夏 2023         推广投入     4333.00  万元  v9_t1_cztr_tgtr
#> 448     新疆 2023         推广投入     2390.16  万元  v9_t1_cztr_tgtr
#> 449     全国 2023     基本建设投入  1240244.34  万元  v9_t1_jstr_jbjs
#> 450     北京 2023     基本建设投入        5.00  万元  v9_t1_jstr_jbjs
#> 451     天津 2023     基本建设投入     7357.00  万元  v9_t1_jstr_jbjs
#> 452     河北 2023     基本建设投入    30555.40  万元  v9_t1_jstr_jbjs
#> 453     山西 2023     基本建设投入     7882.00  万元  v9_t1_jstr_jbjs
#> 454   内蒙古 2023     基本建设投入    85172.00  万元  v9_t1_jstr_jbjs
#> 455     辽宁 2023     基本建设投入     2978.00  万元  v9_t1_jstr_jbjs
#> 456     吉林 2023     基本建设投入    18157.00  万元  v9_t1_jstr_jbjs
#> 457   黑龙江 2023     基本建设投入    11839.00  万元  v9_t1_jstr_jbjs
#> 458     上海 2023     基本建设投入      314.00  万元  v9_t1_jstr_jbjs
#> 459     江苏 2023     基本建设投入   179601.04  万元  v9_t1_jstr_jbjs
#> 460     浙江 2023     基本建设投入   129881.44  万元  v9_t1_jstr_jbjs
#> 461     安徽 2023     基本建设投入    51353.40  万元  v9_t1_jstr_jbjs
#> 462     福建 2023     基本建设投入      203.48  万元  v9_t1_jstr_jbjs
#> 463     江西 2023     基本建设投入    20503.00  万元  v9_t1_jstr_jbjs
#> 464     山东 2023     基本建设投入   193284.26  万元  v9_t1_jstr_jbjs
#> 465     河南 2023     基本建设投入     5547.00  万元  v9_t1_jstr_jbjs
#> 466     湖北 2023     基本建设投入    19030.00  万元  v9_t1_jstr_jbjs
#> 467     湖南 2023     基本建设投入    26346.00  万元  v9_t1_jstr_jbjs
#> 468     广东 2023     基本建设投入    41268.47  万元  v9_t1_jstr_jbjs
#> 469     广西 2023     基本建设投入       77.20  万元  v9_t1_jstr_jbjs
#> 470     海南 2023     基本建设投入     1319.50  万元  v9_t1_jstr_jbjs
#> 471     重庆 2023     基本建设投入   105507.00  万元  v9_t1_jstr_jbjs
#> 472     四川 2023     基本建设投入   227148.00  万元  v9_t1_jstr_jbjs
#> 473     贵州 2023     基本建设投入    24612.00  万元  v9_t1_jstr_jbjs
#> 474     云南 2023     基本建设投入     5926.00  万元  v9_t1_jstr_jbjs
#> 475     西藏 2023     基本建设投入      220.00  万元  v9_t1_jstr_jbjs
#> 476     陕西 2023     基本建设投入     3635.35  万元  v9_t1_jstr_jbjs
#> 477     甘肃 2023     基本建设投入     4831.00  万元  v9_t1_jstr_jbjs
#> 478     青海 2023     基本建设投入    27131.00  万元  v9_t1_jstr_jbjs
#> 479     宁夏 2023     基本建设投入     2701.00  万元  v9_t1_jstr_jbjs
#> 480     新疆 2023     基本建设投入      270.50  万元  v9_t1_jstr_jbjs
#> 481     全国 2023 农业机械购置投入 10160792.41  万元  v9_t1_gztr_jxgz
#> 482     北京 2023 农业机械购置投入    16501.23  万元  v9_t1_gztr_jxgz
#> 483     天津 2023 农业机械购置投入    54363.25  万元  v9_t1_gztr_jxgz
#> 484     河北 2023 农业机械购置投入   636039.78  万元  v9_t1_gztr_jxgz
#> 485     山西 2023 农业机械购置投入   202509.00  万元  v9_t1_gztr_jxgz
#> 486   内蒙古 2023 农业机械购置投入   716631.15  万元  v9_t1_gztr_jxgz
#> 487     辽宁 2023 农业机械购置投入   384319.01  万元  v9_t1_gztr_jxgz
#> 488     吉林 2023 农业机械购置投入   424919.45  万元  v9_t1_gztr_jxgz
#> 489   黑龙江 2023 农业机械购置投入   794509.90  万元  v9_t1_gztr_jxgz
#> 490     上海 2023 农业机械购置投入    20685.00  万元  v9_t1_gztr_jxgz
#> 491     江苏 2023 农业机械购置投入   721021.65  万元  v9_t1_gztr_jxgz
#> 492     浙江 2023 农业机械购置投入   193271.81  万元  v9_t1_gztr_jxgz
#> 493     安徽 2023 农业机械购置投入   793376.22  万元  v9_t1_gztr_jxgz
#> 494     福建 2023 农业机械购置投入   109221.17  万元  v9_t1_gztr_jxgz
#> 495     江西 2023 农业机械购置投入   283044.00  万元  v9_t1_gztr_jxgz
#> 496     山东 2023 农业机械购置投入   820324.96  万元  v9_t1_gztr_jxgz
#> 497     河南 2023 农业机械购置投入   624436.00  万元  v9_t1_gztr_jxgz
#> 498     湖北 2023 农业机械购置投入   736003.40  万元  v9_t1_gztr_jxgz
#> 499     湖南 2023 农业机械购置投入   228260.15  万元  v9_t1_gztr_jxgz
#> 500     广东 2023 农业机械购置投入    77799.31  万元  v9_t1_gztr_jxgz
#> 501     广西 2023 农业机械购置投入   220386.05  万元  v9_t1_gztr_jxgz
#> 502     海南 2023 农业机械购置投入    14046.89  万元  v9_t1_gztr_jxgz
#> 503     重庆 2023 农业机械购置投入    90215.00  万元  v9_t1_gztr_jxgz
#> 504     四川 2023 农业机械购置投入   176705.00  万元  v9_t1_gztr_jxgz
#> 505     贵州 2023 农业机械购置投入    68756.51  万元  v9_t1_gztr_jxgz
#> 506     云南 2023 农业机械购置投入   215393.10  万元  v9_t1_gztr_jxgz
#> 507     西藏 2023 农业机械购置投入    42390.82  万元  v9_t1_gztr_jxgz
#> 508     陕西 2023 农业机械购置投入   162953.54  万元  v9_t1_gztr_jxgz
#> 509     甘肃 2023 农业机械购置投入   270174.30  万元  v9_t1_gztr_jxgz
#> 510     青海 2023 农业机械购置投入    37011.00  万元  v9_t1_gztr_jxgz
#> 511     宁夏 2023 农业机械购置投入   110146.00  万元  v9_t1_gztr_jxgz
#> 512     新疆 2023 农业机械购置投入   519063.09  万元  v9_t1_gztr_jxgz
#> 513     全国 2023     农机服务收入 48290124.50  万元 v9_t1_zsr_njfwsr
#> 514     北京 2023     农机服务收入    17040.40  万元 v9_t1_zsr_njfwsr
#> 515     天津 2023     农机服务收入    93807.00  万元 v9_t1_zsr_njfwsr
#> 516     河北 2023     农机服务收入  1964047.20  万元 v9_t1_zsr_njfwsr
#> 517     山西 2023     农机服务收入   870216.00  万元 v9_t1_zsr_njfwsr
#> 518   内蒙古 2023     农机服务收入  1860960.09  万元 v9_t1_zsr_njfwsr
#> 519     辽宁 2023     农机服务收入  1123882.90  万元 v9_t1_zsr_njfwsr
#> 520     吉林 2023     农机服务收入  1770974.00  万元 v9_t1_zsr_njfwsr
#> 521   黑龙江 2023     农机服务收入  1776831.00  万元 v9_t1_zsr_njfwsr
#> 522     上海 2023     农机服务收入    35561.00  万元 v9_t1_zsr_njfwsr
#> 523     江苏 2023     农机服务收入  4334624.45  万元 v9_t1_zsr_njfwsr
#> 524     浙江 2023     农机服务收入   791236.13  万元 v9_t1_zsr_njfwsr
#> 525     安徽 2023     农机服务收入  5555295.53  万元 v9_t1_zsr_njfwsr
#> 526     福建 2023     农机服务收入   634188.50  万元 v9_t1_zsr_njfwsr
#> 527     江西 2023     农机服务收入  1791193.00  万元 v9_t1_zsr_njfwsr
#> 528     山东 2023     农机服务收入  4392255.67  万元 v9_t1_zsr_njfwsr
#> 529     河南 2023     农机服务收入  2807993.00  万元 v9_t1_zsr_njfwsr
#> 530     湖北 2023     农机服务收入  2436913.00  万元 v9_t1_zsr_njfwsr
#> 531     湖南 2023     农机服务收入  3649823.00  万元 v9_t1_zsr_njfwsr
#> 532     广东 2023     农机服务收入  1437256.59  万元 v9_t1_zsr_njfwsr
#> 533     广西 2023     农机服务收入  2948979.05  万元 v9_t1_zsr_njfwsr
#> 534     海南 2023     农机服务收入   332462.00  万元 v9_t1_zsr_njfwsr
#> 535     重庆 2023     农机服务收入   379082.00  万元 v9_t1_zsr_njfwsr
#> 536     四川 2023     农机服务收入  1756905.00  万元 v9_t1_zsr_njfwsr
#> 537     贵州 2023     农机服务收入   270812.88  万元 v9_t1_zsr_njfwsr
#> 538     云南 2023     农机服务收入   940986.91  万元 v9_t1_zsr_njfwsr
#> 539     西藏 2023     农机服务收入    15277.02  万元 v9_t1_zsr_njfwsr
#> 540     陕西 2023     农机服务收入  1072764.41  万元 v9_t1_zsr_njfwsr
#> 541     甘肃 2023     农机服务收入   946504.00  万元 v9_t1_zsr_njfwsr
#> 542     青海 2023     农机服务收入    85669.00  万元 v9_t1_zsr_njfwsr
#> 543     宁夏 2023     农机服务收入   261091.00  万元 v9_t1_zsr_njfwsr
#> 544     新疆 2023     农机服务收入  1559165.77  万元 v9_t1_zsr_njfwsr