Helper Function to match and get province of target institution

Helper Function to match and get province of target institution

Usage

get_province_of_institution(df, target_institution, target_province)

Arguments

df,

data frame, contains the target institution column

target_institution,

character, the column name of target institution

target_province,

character, the target output column name for province variable.

Value

out

Examples

library(magrittr)
utils::data("PubObsStation",package = "techme")
tbl_read <- PubObsStation %>%
  dplyr::select(-province)

tbl_out <- get_province_of_institution(df = tbl_read,
  target_institution ="institution",
  target_province = "province")

head(tbl_out)
#> # A tibble: 6 × 6
#>   officer  year index name                       institution            province
#>   <chr>   <dbl> <chr> <chr>                      <chr>                  <chr>   
#> 1 MOA      2018 1     国家土壤质量德州观测实验站 中国农业科学院农业资源与农业区划研究所…… 北京    
#> 2 MOA      2018 2     国家土壤质量祁阳观测实验站 中国农业科学院农业资源与农业区划研究所…… 北京    
#> 3 MOA      2018 3     国家土壤质量昌平观测实验站 中国农业科学院农业资源与农业区划研究所…… 北京    
#> 4 MOA      2018 4     国家土壤质量洛龙观测实验站 中国农业科学院农业资源与农业区划研究所…… 北京    
#> 5 MOA      2018 5     国家植物保护甘谷观测实验站 中国农业科学院植物保护研究所…… 北京    
#> 6 MOA      2018 6     国家种质资源管城观测实验站 中国农业科学院郑州果树研究所…… 河南